refactor: split monolithic backends into category files
Vulkan (2,822→1,216 lines in main file):
- vk_peak.cpp: engine room only (VulkanDevice, vkPeak core, runAll,
  runComputeKernel, inventory)
- compute_float.cpp: SP, HP, DP, MP, BF16
- compute_int.cpp: Int32, Int8DP, Int4Packed
- coopmat.cpp: cooperative matrix (tensor-core) umbrella
- global_bandwidth.cpp, local_bandwidth.cpp, image_bandwidth.cpp,
  transfer_bandwidth.cpp: bandwidth tests
- atomic_throughput.cpp: int + fp atomics
- kernel_latency.cpp: kernel launch latency
- Moved targetVulkanGlobalThreads to vk_peak.h as static inline

CUDA (1,526→585 lines in main file):
- cuda_peak.cpp: engine room only (CudaDevice, CudaPeak core, runAll,
  runComputeKernel, inventory)
- compute_float.cpp, compute_int.cpp: compute wrappers
- wmma.cpp: WMMA/MMA tensor-core umbrella (extracted from main)
- Individual bandwidth/latency/atomic files
- cuda_blas.cpp unchanged (already separate)

Metal (1,143→502 lines in main file):
- mtl_peak.mm: engine room + helpers (mtlGetLibrary, mtlGetPipeline,
  mtlRunDispatches) + inventory
- mtl_internal.h: shared internal header with ObjC types, pimpl defs,
  and helper declarations — included by all .mm category files
- compute_float.mm, compute_int.mm: compute wrappers
- simdgroup.mm: simdgroup matrix multiply
- Individual bandwidth/latency/atomic files
- mtl_blas.mm unchanged (already separate)

OpenCL: already well-organized; only fixed kernel-latency skip-path
display string for consistency.

AGENTS.md updated for Vulkan, CUDA, and Metal.

All backends now share the same category-based layout so contributors
know exactly which file to touch when adding a new benchmark.

Fix: standalone tests now include unit in display name:
  'Atomic throughput' → 'Atomic throughput (GOPS)' / '(GFLOPS)'
  'Global memory bandwidth' → 'Global memory bandwidth (GBPS)'
  'Kernel launch latency' → 'Kernel launch latency (us)'
  'MPS GEMM peak' → 'MPS GEMM peak (TFLOPS)' / '(TOPS)'
