+++ src/cuda/AGENTS.md
@@ -9,7 +9,13 @@ CUDA kernels (in `cuda_kernels/`) compiled via NVRTC at runtime.  Built as
 - Looking for the main class? → `cuda_peak.cpp`
 - Looking for driver init / device enumeration? → `cuda_peak.cpp` (`initDriver`)
 - Looking for the unified compute kernel runner? → `cuda_peak.cpp` (`runComputeKernel`)
+- Looking for FP compute benchmarks? → `compute_float.cpp`
+- Looking for int compute benchmarks? → `compute_int.cpp`
+- Looking for WMMA/MMA benchmarks? → `wmma.cpp`
 - Looking for cuBLAS benchmarks? → `cuda_blas.cpp`
+- Looking for bandwidth benchmarks? → `global_bandwidth.cpp`, `local_bandwidth.cpp`, `image_bandwidth.cpp`, `transfer_bandwidth.cpp`
+- Looking for atomic benchmarks? → `atomic_throughput.cpp`
+- Looking for kernel latency? → `kernel_latency.cpp`
 - Looking for .cu kernel sources? → `cuda_kernels/*.cu`
 - Looking for kernel embedding logic? → `cmake/EmbedCudaKernels.cmake`
 
@@ -17,14 +23,23 @@ CUDA kernels (in `cuda_kernels/`) compiled via NVRTC at runtime.  Built as
 
 | File | Purpose |
 |------|---------|
-| `cuda_peak.cpp` | `CudaPeak` class: constructor, `applyOptions()`, `runAll()`, all benchmarks |
-| `cuda_blas.cpp` | `runCublas()` — cuBLASLt tensor-core GEMM benchmarks |
+| `cuda_peak.cpp` | `CudaPeak` + `CudaDevice` classes: ctor, `applyOptions()`, `runAll()`, `runComputeKernel()`, `run_kernel()`, inventory |
+| `compute_float.cpp` | `runComputeSP`, `runComputeHP`, `runComputeDP`, `runComputeMP`, `runComputeBF16` |
+| `compute_int.cpp` | `runComputeInt32`, `runComputeInt8DP`, `runComputeInt4Packed` |
+| `wmma.cpp` | `runWmma` — WMMA/MMA tensor-core umbrella |
+| `cuda_blas.cpp` | `runCublas` — cuBLASLt tensor-core GEMM benchmarks |
+| `global_bandwidth.cpp` | `runGlobalBandwidth` |
+| `local_bandwidth.cpp` | `runLocalBandwidth` |
+| `image_bandwidth.cpp` | `runImageBandwidth` |
+| `transfer_bandwidth.cpp` | `runTransferBandwidth` |
+| `atomic_throughput.cpp` | `runAtomicThroughput` |
+| `kernel_latency.cpp` | `runKernelLatency` |
 | `cuda_kernels/` | CUDA kernel sources (`.cu`) embedded as C++ string literals |
 | `cmake/EmbedCudaKernels.cmake` | `embed_cuda_kernels()` — .cu → C++ raw-string arrays |
 
 ## When You Change This Directory
 
-- If you add a new benchmark → update `CMakeLists.txt` (kernel list) + this file.
+- If you add a new benchmark → add it to the appropriate category file + update `CMakeLists.txt` + this file.
 - If you add a new `.cu` kernel → add to `CLPEAK_CUDA_KERNELS` in `CMakeLists.txt`.
 - If you change `CudaPeak` interface → update `include/cuda/cuda_peak.h`.
 - If you change the NVRTC include path → check `CLPEAK_CUDA_INCLUDE_DIR` compile definition.
