+++ src/cuda/cuda_peak.cpp
@@ -517,953 +517,12 @@ int CudaPeak::runComputeKernel(CudaDevice &dev, benchmark_config_t &cfg,
 }
 
 // ---------------------------------------------------------------------------
-// Concrete compute benchmarks.
+// Benchmark methods live in separate category files:
+//   compute_float.cpp    compute_int.cpp    wmma.cpp
+//   global_bandwidth.cpp local_bandwidth.cpp image_bandwidth.cpp
+//   transfer_bandwidth.cpp kernel_latency.cpp atomic_throughput.cpp
+//   cuda_blas.cpp (cuBLAS GEMM)
 // ---------------------------------------------------------------------------
-
-int CudaPeak::runComputeSP(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  float A = 1.3f;
-  cuda_compute_desc_t d = {};
-  d.title = "Single-precision compute (GFLOPS)";
-  d.resultTag = "single_precision_compute";
-  d.unit = "gflops";
-  d.metricLabel = "float";
-  d.kernelName = "compute_sp";
-  d.src = cuda_kernels::compute_sp_src;
-  d.srcName = cuda_kernels::compute_sp_name;
-  d.workPerWI = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize = sizeof(float);
-  d.scalarArg = &A;
-  d.scalarSize = sizeof(A);
-  return runComputeKernel(dev, cfg, d);
-}
-
-int CudaPeak::runComputeHP(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  static const cuda_compute_variant_t variants[] = {
-      {"half", "compute_hp", cuda_kernels::compute_hp_src, cuda_kernels::compute_hp_name},
-      {"half2", "compute_hp2", cuda_kernels::compute_hp_src, cuda_kernels::compute_hp_name},
-  };
-  float A = 1.3f;
-  cuda_compute_desc_t d = {};
-  d.title = "Half-precision compute (GFLOPS)";
-  d.resultTag = "half_precision_compute";
-  d.unit = "gflops";
-  d.variants = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize = sizeof(float); // 32-bit slot per thread; we store the reduced fp32 result
-  d.scalarArg = &A;
-  d.scalarSize = sizeof(A);
-  d.skip = !dev.info.fp16Supported;
-  d.skipMsg = "fp16 not supported on this compute capability! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-
-int CudaPeak::runComputeDP(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  double A = 1.3;
-  cuda_compute_desc_t d = {};
-  d.title = "Double-precision compute (GFLOPS)";
-  d.resultTag = "double_precision_compute";
-  d.unit = "gflops";
-  d.metricLabel = "double";
-  d.kernelName = "compute_dp";
-  d.src = cuda_kernels::compute_dp_src;
-  d.srcName = cuda_kernels::compute_dp_name;
-  d.workPerWI = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize = sizeof(double);
-  d.scalarArg = &A;
-  d.scalarSize = sizeof(A);
-  return runComputeKernel(dev, cfg, d);
-}
-
-int CudaPeak::runComputeMP(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  // Single variant: NVIDIA shader-core fp16xfp16+fp32 issues at FP32 rate.
-  // The packed (HFMA2) path is fp16xfp16+fp16 -- that's compute_hp2, not MP.
-  float A = 1.3f;
-  cuda_compute_desc_t d = {};
-  d.title = "Mixed-precision compute fp16xfp16+fp32 (GFLOPS)";
-  d.resultTag = "mixed_precision_compute";
-  d.unit = "gflops";
-  d.metricLabel = "mp";
-  d.kernelName = "compute_mp";
-  d.src = cuda_kernels::compute_mp_src;
-  d.srcName = cuda_kernels::compute_mp_name;
-  d.workPerWI = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize = sizeof(float);
-  d.scalarArg = &A;
-  d.scalarSize = sizeof(A);
-  d.skip = !dev.info.fp16Supported;
-  d.skipMsg = "fp16 not supported on this compute capability! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-
-int CudaPeak::runComputeBF16(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  // Single variant: shader-core bf16xbf16+fp32 issues at FP32 rate on
-  // Ampere+.  Packed BF16 is reachable through tensor cores (wmma), not
-  // an SFU-style packed shader instruction, so a bf16_2 variant wouldn't
-  // be a different code path.
-  float A = 1.3f;
-  cuda_compute_desc_t d = {};
-  d.title = "BF16 compute bf16xbf16+fp32 (GFLOPS)";
-  d.resultTag = "bfloat16_compute";
-  d.unit = "gflops";
-  d.metricLabel = "bf16";
-  d.kernelName = "compute_bf16";
-  d.src = cuda_kernels::compute_bf16_src;
-  d.srcName = cuda_kernels::compute_bf16_name;
-  d.workPerWI = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize = sizeof(float);
-  d.scalarArg = &A;
-  d.scalarSize = sizeof(A);
-  d.skip = !dev.info.bf16Supported;
-  d.skipMsg = "bf16 requires sm_80 or newer (Ampere+)! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-
-int CudaPeak::runComputeInt32(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  // Scalar 32-bit integer IMAD chain throughput.  Distinct shader-core
-  // path from __dp4a (compute_int8_dp) and the int4 emulation; reported
-  // in GOPS.
-  int A = 3;
-  cuda_compute_desc_t d = {};
-  d.title = "Integer compute (32-bit IMAD) (GOPS)";
-  d.resultTag = "integer_compute";
-  d.unit = "gops";
-  d.metricLabel = "int";
-  d.kernelName = "compute_int32";
-  d.src = cuda_kernels::compute_int32_src;
-  d.srcName = cuda_kernels::compute_int32_name;
-  d.workPerWI = COMPUTE_FP_WORK_PER_WI; // 4096 ops/thread (same scaling)
-  d.elemSize = sizeof(int);
-  d.scalarArg = &A;
-  d.scalarSize = sizeof(A);
-  return runComputeKernel(dev, cfg, d);
-}
-
-int CudaPeak::runComputeInt8DP(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  static const cuda_compute_variant_t variants[] = {
-      {"int8_dp", "compute_int8_dp", cuda_kernels::compute_int8_dp_src, cuda_kernels::compute_int8_dp_name},
-      {"int8_dp2", "compute_int8_dp2", cuda_kernels::compute_int8_dp_src, cuda_kernels::compute_int8_dp_name},
-      {"int8_dp4", "compute_int8_dp4", cuda_kernels::compute_int8_dp_src, cuda_kernels::compute_int8_dp_name},
-      {"int8_dp8", "compute_int8_dp8", cuda_kernels::compute_int8_dp_src, cuda_kernels::compute_int8_dp_name},
-  };
-  int A = 4;
-  cuda_compute_desc_t d = {};
-  d.title = "INT8 dot-product compute (__dp4a) (GOPS)";
-  d.resultTag = "integer_compute_int8_dp";
-  d.unit = "gops";
-  d.variants = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI = COMPUTE_INT8_DP_WORK_PER_WI;
-  d.elemSize = sizeof(int);
-  d.scalarArg = &A;
-  d.scalarSize = sizeof(A);
-  d.skip = !dev.info.dp4aSupported;
-  d.skipMsg = "__dp4a requires sm_61 or newer (Pascal+)! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-
-int CudaPeak::runComputeInt4Packed(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  int A = 3;
-  cuda_compute_desc_t d = {};
-  d.title = "Packed INT4 compute (emulated) (GOPS)";
-  d.resultTag = "int4_packed_compute";
-  d.unit = "gops";
-  d.metricLabel = "int4_packed";
-  d.kernelName = "compute_int4_packed";
-  d.src = cuda_kernels::compute_int4_packed_src;
-  d.srcName = cuda_kernels::compute_int4_packed_name;
-  d.workPerWI = COMPUTE_INT4_PACKED_WORK_PER_WI;
-  d.elemSize = sizeof(int);
-  d.scalarArg = &A;
-  d.scalarSize = sizeof(A);
-  return runComputeKernel(dev, cfg, d);
-}
-
-// ---------------------------------------------------------------------------
-// WMMA + FP8 mma.sync umbrella -- mirrors vkPeak::runCoopMatrix.
-// ---------------------------------------------------------------------------
-
-int CudaPeak::runWmma(CudaDevice &dev, benchmark_config_t &cfg, Category category)
-{
-  // Shared geometry: one warp (32 threads) per block, m16n16k16 tile per
-  // wmma fragment, 256 outer iters → COOPMAT_WORK_PER_WI per thread.
-  const uint32_t warp = 32;
-  const uint32_t outElems = 16 * 16; // M*N
-
-  // ---------------------------------------------------------------------
-  // FP cluster -- each variant opens its own <wmma_*> group with the
-  // proper unit attribute via runComputeKernel; no umbrella tag here
-  // (depth-5 nesting under one would break the v2 logger shim).
-  // ---------------------------------------------------------------------
-  if (category == Category::FpCompute)
-  {
-    // FP16 WMMA
-    {
-      float A = 1.3f;
-      cuda_compute_desc_t d = {};
-      d.title = "WMMA fp16xfp16+fp32 16x16x16 (TFLOPS)";
-      d.resultTag = "wmma_fp16";
-      d.unit = "tflops";
-      d.unitDivider = 1e12;
-      d.metricLabel = "wmma_fp16";
-      d.kernelName = "wmma_fp16";
-      d.src = cuda_kernels::wmma_fp16_src;
-      d.srcName = cuda_kernels::wmma_fp16_name;
-      d.workPerWI = COOPMAT_WORK_PER_WI * 4; // 4 parallel chains per kernel
-      d.elemSize = sizeof(float);
-      d.blockSize = warp;
-      d.outElemsPerBlock = outElems;
-      d.scalarArg = &A;
-      d.scalarSize = sizeof(A);
-      d.skip = !dev.info.wmmaSupported;
-      d.skipMsg = "WMMA requires sm_70 or newer (Volta+)! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-    // BF16 WMMA
-    {
-      float A = 1.3f;
-      cuda_compute_desc_t d = {};
-      d.title = "WMMA bf16xbf16+fp32 16x16x16 (TFLOPS)";
-      d.resultTag = "wmma_bf16";
-      d.unit = "tflops";
-      d.unitDivider = 1e12;
-      d.metricLabel = "wmma_bf16";
-      d.kernelName = "wmma_bf16";
-      d.src = cuda_kernels::wmma_bf16_src;
-      d.srcName = cuda_kernels::wmma_bf16_name;
-      d.workPerWI = COOPMAT_WORK_PER_WI * 4;
-      d.elemSize = sizeof(float);
-      d.blockSize = warp;
-      d.outElemsPerBlock = outElems;
-      d.scalarArg = &A;
-      d.scalarSize = sizeof(A);
-      d.skip = !dev.info.wmmaSupported || !dev.info.bf16Supported;
-      d.skipMsg = "bf16 WMMA requires sm_80 or newer (Ampere+)! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-    // TF32 WMMA m16n16k8 -- Ampere+
-    {
-      float A = 1.3f;
-      cuda_compute_desc_t d = {};
-      d.title = "WMMA tf32xtf32+fp32 16x16x8 (TFLOPS)";
-      d.resultTag = "wmma_tf32";
-      d.unit = "tflops";
-      d.unitDivider = 1e12;
-      d.metricLabel = "wmma_tf32";
-      d.kernelName = "wmma_tf32";
-      d.src = cuda_kernels::wmma_tf32_src;
-      d.srcName = cuda_kernels::wmma_tf32_name;
-      d.workPerWI = COOPMAT_WORK_PER_WI * 2; // m16n16k8 = half the K of fp16
-      d.elemSize = sizeof(float);
-      d.blockSize = warp;
-      d.outElemsPerBlock = outElems;
-      d.scalarArg = &A;
-      d.scalarSize = sizeof(A);
-      d.skip = !dev.info.wmmaSupported || !dev.info.tf32GemmSupported;
-      d.skipMsg = "TF32 WMMA requires sm_80 or newer (Ampere+)! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-    // FP64 WMMA m8n8k4 -- Ampere+ DP tensor cores
-    {
-      double A = 1.3;
-      cuda_compute_desc_t d = {};
-      d.title = "WMMA fp64xfp64+fp64 8x8x4 (TFLOPS)";
-      d.resultTag = "wmma_fp64";
-      d.unit = "tflops";
-      d.unitDivider = 1e12;
-      d.metricLabel = "wmma_fp64";
-      d.kernelName = "wmma_fp64";
-      d.src = cuda_kernels::wmma_fp64_src;
-      d.srcName = cuda_kernels::wmma_fp64_name;
-      d.workPerWI = COOPMAT_WORK_PER_WI; // 1024 outer iters bring this to par
-      d.elemSize = sizeof(double);
-      d.blockSize = warp;
-      d.outElemsPerBlock = 8 * 8;
-      d.scalarArg = &A;
-      d.scalarSize = sizeof(A);
-      d.skip = !dev.info.wmmaSupported || !dev.info.dpTensorSupported;
-      d.skipMsg = "FP64 WMMA requires sm_80 or newer (Ampere+)! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-    // FP8 mma.sync E4M3 (PTX) - sm_89+
-    {
-      float A = 1.3f;
-      cuda_compute_desc_t d = {};
-      d.title = "FP8(E4M3) mma.sync m16n8k32+fp32 (TFLOPS)";
-      d.resultTag = "wmma_fp8_e4m3";
-      d.unit = "tflops";
-      d.unitDivider = 1e12;
-      d.metricLabel = "fp8_e4m3";
-      d.kernelName = "wmma_fp8_e4m3";
-      d.src = cuda_kernels::wmma_fp8_e4m3_src;
-      d.srcName = cuda_kernels::wmma_fp8_e4m3_name;
-      d.workPerWI = COOPMAT_WORK_PER_WI * 8; // 8 parallel chains for FP8
-      d.elemSize = sizeof(float);
-      d.blockSize = warp;
-      d.outElemsPerBlock = 16 * 8;
-      d.scalarArg = &A;
-      d.scalarSize = sizeof(A);
-      d.skip = !dev.info.wmmaSupported || !dev.info.fp8MmaSupported;
-      d.skipMsg = "FP8 mma.sync requires sm_89 or newer (Ada/Hopper+)! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-    // FP8 mma.sync E5M2 (PTX) - sm_89+
-    {
-      float A = 1.3f;
-      cuda_compute_desc_t d = {};
-      d.title = "FP8(E5M2) mma.sync m16n8k32+fp32 (TFLOPS)";
-      d.resultTag = "wmma_fp8_e5m2";
-      d.unit = "tflops";
-      d.unitDivider = 1e12;
-      d.metricLabel = "fp8_e5m2";
-      d.kernelName = "wmma_fp8_e5m2";
-      d.src = cuda_kernels::wmma_fp8_e5m2_src;
-      d.srcName = cuda_kernels::wmma_fp8_e5m2_name;
-      d.workPerWI = COOPMAT_WORK_PER_WI * 8;
-      d.elemSize = sizeof(float);
-      d.blockSize = warp;
-      d.outElemsPerBlock = 16 * 8;
-      d.scalarArg = &A;
-      d.scalarSize = sizeof(A);
-      d.skip = !dev.info.wmmaSupported || !dev.info.fp8MmaSupported;
-      d.skipMsg = "FP8 mma.sync requires sm_89 or newer (Ada/Hopper+)! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-    // FP4 mma.sync E2M1 (PTX) - Blackwell sm_120a+
-    {
-      float A = 1.3f;
-      std::stringstream archOpt;
-      archOpt << "--gpu-architecture=sm_" << dev.info.major << dev.info.minor << "a";
-      std::string archOptStr = archOpt.str();
-      const char *fp4Opts[] = {archOptStr.c_str()};
-      cuda_compute_desc_t d = {};
-      d.title = "FP4(E2M1) mma.sync m16n8k32+fp32 (TFLOPS)";
-      d.resultTag = "wmma_fp4_e2m1";
-      d.unit = "tflops";
-      d.unitDivider = 1e12;
-      d.metricLabel = "fp4_e2m1";
-      d.kernelName = "wmma_fp4_e2m1";
-      d.src = cuda_kernels::wmma_fp4_e2m1_src;
-      d.srcName = cuda_kernels::wmma_fp4_e2m1_name;
-      d.workPerWI = COOPMAT_WORK_PER_WI * 8;
-      d.elemSize = sizeof(float);
-      d.blockSize = warp;
-      d.outElemsPerBlock = 16 * 8;
-      d.scalarArg = &A;
-      d.scalarSize = sizeof(A);
-      d.skip = !dev.info.wmmaSupported || !dev.info.fp4MmaSupported;
-      d.skipMsg = "FP4 mma.sync requires Blackwell sm_120a or newer! Skipped";
-      d.extraNvrtcOpts = fp4Opts;
-      d.numExtraNvrtcOpts = 1;
-      runComputeKernel(dev, cfg, d);
-    }
-    // MXFP4 mma.sync E2M1 + UE8M0 block scale (PTX) - Blackwell sm_120a+
-    {
-      float A = 1.3f;
-      std::stringstream archOpt;
-      archOpt << "--gpu-architecture=sm_" << dev.info.major << dev.info.minor << "a";
-      std::string archOptStr = archOpt.str();
-      const char *fp4Opts[] = {archOptStr.c_str()};
-      cuda_compute_desc_t d = {};
-      d.title = "MXFP4(E2M1) mma.sync m16n8k64+fp32 (TFLOPS)";
-      d.resultTag = "wmma_mxf4_e2m1";
-      d.unit = "tflops";
-      d.unitDivider = 1e12;
-      d.metricLabel = "mxf4_e2m1";
-      d.kernelName = "wmma_mxf4_e2m1";
-      d.src = cuda_kernels::wmma_mxf4_e2m1_src;
-      d.srcName = cuda_kernels::wmma_mxf4_e2m1_name;
-      d.workPerWI = COOPMAT_WORK_PER_WI * 16;
-      d.elemSize = sizeof(float);
-      d.blockSize = warp;
-      d.outElemsPerBlock = 16 * 8;
-      d.scalarArg = &A;
-      d.scalarSize = sizeof(A);
-      d.skip = !dev.info.wmmaSupported || !dev.info.fp4MmaSupported;
-      d.skipMsg = "MXFP4 mma.sync requires Blackwell sm_120a or newer! Skipped";
-      d.extraNvrtcOpts = fp4Opts;
-      d.numExtraNvrtcOpts = 1;
-      runComputeKernel(dev, cfg, d);
-    }
-  }
-
-  // ---------------------------------------------------------------------
-  // Integer / binary cluster -- each variant opens its own <wmma_*> group
-  // with the proper unit attribute via runComputeKernel; no umbrella tag.
-  // The fp/int split is preserved by the per-variant unit -> category
-  // derivation in the dump pipeline.
-  // ---------------------------------------------------------------------
-
-  if (category != Category::IntCompute)
-    return 0;
-
-  // INT8 WMMA
-  {
-    int A = 3;
-    cuda_compute_desc_t d = {};
-    d.title = "WMMA int8xint8+int32 16x16x16 (TOPS)";
-    d.resultTag = "wmma_int8";
-    d.unit = "tops";
-    d.unitDivider = 1e12;
-    d.metricLabel = "wmma_int8";
-    d.kernelName = "wmma_int8";
-    d.src = cuda_kernels::wmma_int8_src;
-    d.srcName = cuda_kernels::wmma_int8_name;
-    d.workPerWI = COOPMAT_WORK_PER_WI * 4;
-    d.elemSize = sizeof(int);
-    d.blockSize = warp;
-    d.outElemsPerBlock = outElems;
-    d.scalarArg = &A;
-    d.scalarSize = sizeof(A);
-    d.skip = !dev.info.wmmaSupported || !dev.info.wmmaInt8Supported;
-    d.skipMsg = "INT8 WMMA requires sm_72 or newer (Turing+)! Skipped";
-    runComputeKernel(dev, cfg, d);
-  }
-  // INT8 mma.sync K=32 (NVIDIA-native tile via inline PTX)
-  {
-    int A = 3;
-    cuda_compute_desc_t d = {};
-    d.title = "INT8 mma.sync m16n8k32+int32 (TOPS)";
-    d.resultTag = "wmma_int8_k32";
-    d.unit = "tops";
-    d.unitDivider = 1e12;
-    d.metricLabel = "int8_k32";
-    d.kernelName = "wmma_int8_k32";
-    d.src = cuda_kernels::wmma_int8_k32_src;
-    d.srcName = cuda_kernels::wmma_int8_k32_name;
-    d.workPerWI = COOPMAT_WORK_PER_WI * 4;
-    d.elemSize = sizeof(int);
-    d.blockSize = warp;
-    d.outElemsPerBlock = 16 * 8;
-    d.scalarArg = &A;
-    d.scalarSize = sizeof(A);
-    d.skip = !dev.info.wmmaSupported || !dev.info.wmmaInt8Supported;
-    d.skipMsg = "INT8 mma.sync K=32 requires sm_72 or newer (Turing+)! Skipped";
-    runComputeKernel(dev, cfg, d);
-  }
-  // INT8 mma.sp 2:4 structured sparsity m16n8k32 -- sm_80+
-  {
-    int A = 3;
-    cuda_compute_desc_t d = {};
-    d.title = "INT8 mma.sp 2:4 sparsity m16n8k32+int32 (TOPS)";
-    d.resultTag = "wmma_int8_sparse";
-    d.unit = "tops";
-    d.unitDivider = 1e12;
-    d.metricLabel = "int8_sparse";
-    d.kernelName = "wmma_int8_sparse";
-    d.src = cuda_kernels::wmma_int8_sparse_src;
-    d.srcName = cuda_kernels::wmma_int8_sparse_name;
-    d.workPerWI = COOPMAT_WORK_PER_WI * 4;
-    d.elemSize = sizeof(int);
-    d.blockSize = warp;
-    d.outElemsPerBlock = 16 * 8;
-    d.scalarArg = &A;
-    d.scalarSize = sizeof(A);
-    d.skip = !dev.info.wmmaSupported || !dev.info.int8MmaSparseSupported;
-    d.skipMsg = "INT8 mma.sp 2:4 sparsity requires sm_80 or newer (Ampere+)! Skipped";
-    runComputeKernel(dev, cfg, d);
-  }
-  // INT4 mma.sync m8n8k32 -- sm_75..sm_89
-  {
-    int A = 3;
-    cuda_compute_desc_t d = {};
-    d.title = "INT4 mma.sync m8n8k32+int32 (TOPS)";
-    d.resultTag = "wmma_int4";
-    d.unit = "tops";
-    d.unitDivider = 1e12;
-    d.metricLabel = "int4";
-    d.kernelName = "wmma_int4";
-    d.src = cuda_kernels::wmma_int4_src;
-    d.srcName = cuda_kernels::wmma_int4_name;
-    d.workPerWI = COOPMAT_WORK_PER_WI * 2; // 256 outer * 4 chains * 8*8*32*2 / 32
-    d.elemSize = sizeof(int);
-    d.blockSize = warp;
-    d.outElemsPerBlock = 8 * 8;
-    d.scalarArg = &A;
-    d.scalarSize = sizeof(A);
-    d.skip = !dev.info.wmmaSupported || !dev.info.int4MmaSupported;
-    d.skipMsg = "INT4 mma.sync requires sm_75..sm_89 (Turing/Ampere/Ada)! Skipped";
-    runComputeKernel(dev, cfg, d);
-  }
-  // BMMA b1 mma.sync m8n8k128 (XOR-popc) -- sm_75+
-  {
-    int A = 3;
-    cuda_compute_desc_t d = {};
-    d.title = "BMMA b1 mma.sync m8n8k128+int32 xor.popc (TOPS)";
-    d.resultTag = "wmma_bmma_b1";
-    d.unit = "tops";
-    d.unitDivider = 1e12;
-    d.metricLabel = "bmma_b1";
-    d.kernelName = "wmma_bmma_b1";
-    d.src = cuda_kernels::wmma_bmma_b1_src;
-    d.srcName = cuda_kernels::wmma_bmma_b1_name;
-    d.workPerWI = COOPMAT_WORK_PER_WI * 8; // 256 outer * 4 chains * 8*8*128*2 / 32
-    d.elemSize = sizeof(int);
-    d.blockSize = warp;
-    d.outElemsPerBlock = 8 * 8;
-    d.scalarArg = &A;
-    d.scalarSize = sizeof(A);
-    d.skip = !dev.info.wmmaSupported || !dev.info.bmmaSupported;
-    d.skipMsg = "BMMA b1 requires sm_75 or newer (Turing+)! Skipped";
-    runComputeKernel(dev, cfg, d);
-  }
-
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Global bandwidth (CUDA)
-// ---------------------------------------------------------------------------
-
-int CudaPeak::runGlobalBandwidth(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  const uint32_t blockSize = 256;
-
-  uint64_t maxItems = dev.info.totalGlobalMem / sizeof(float) / 4; // input+output, plus margin
-  uint64_t numItems = (maxItems / (blockSize * FETCH_PER_WI)) * (blockSize * FETCH_PER_WI);
-  if (numItems > cfg.globalBWMaxSize / sizeof(float))
-    numItems = (cfg.globalBWMaxSize / sizeof(float) / (blockSize * FETCH_PER_WI)) * (blockSize * FETCH_PER_WI);
-
-  uint32_t numBlocks = (uint32_t)(numItems / FETCH_PER_WI / blockSize);
-  if (numBlocks == 0)
-    numBlocks = 1;
-
-  auto test = currentDeviceScope->beginTest(
-    {"global_memory_bandwidth", "Global memory bandwidth (GBPS)", "gbps"});
-
-  CUdeviceptr inBuf = 0, outBuf = 0;
-  if (cuMemAlloc(&inBuf, numItems * sizeof(float)) != CUDA_SUCCESS ||
-      cuMemAlloc(&outBuf, numItems * sizeof(float)) != CUDA_SUCCESS)
-  {
-    const char *labels[] = {"float", "float2", "float4"};
-    for (int i = 0; i < 3; i++)
-      test.skip(labels[i], ResultStatus::Error, "Failed to allocate buffers");
-    if (inBuf)
-      cuMemFree(inBuf);
-    return -1;
-  }
-  // Touch input so we measure DRAM not zero-page.
-  cuMemsetD32(inBuf, 0x3f800000u, numItems);
-
-  // Each variant takes input typed as floatN* so element index strides by
-  // V floats per iteration -- matching the OpenCL backend.  numBlocks
-  // shrinks accordingly so the total bytes touched remain the same.
-  struct Variant
-  {
-    const char *label;
-    const char *kernelName;
-    uint32_t width;
-  };
-  static const Variant variants[] = {
-      {"float   ", "global_bandwidth_v1", 1},
-      {"float2  ", "global_bandwidth_v2", 2},
-      {"float4  ", "global_bandwidth_v4", 4},
-  };
-
-  for (const auto &v : variants)
-  {
-    std::string key(v.label);
-    while (!key.empty() && key.back() == ' ')
-      key.pop_back();
-
-    CUfunction fn;
-    if (!dev.getKernel(cuda_kernels::global_bandwidth_src,
-                       cuda_kernels::global_bandwidth_name,
-                       v.kernelName, fn))
-    {
-      test.skip(key, ResultStatus::Error, "Kernel compile failed");
-      continue;
-    }
-
-    uint64_t blocks = numItems / FETCH_PER_WI / v.width / blockSize;
-    if (blocks == 0)
-      blocks = 1;
-    uint32_t blocksU = (uint32_t)blocks;
-
-    void *args[2] = {&inBuf, &outBuf};
-    float us = runKernel(dev, fn, blocksU, blockSize, args,
-                         cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-    double bytes = (double)blocksU * blockSize * FETCH_PER_WI * v.width * sizeof(float);
-    float gbps = (float)(bytes / us / 1e3);
-    test.emit(key, gbps);
-  }
-
-  cuMemFree(inBuf);
-  cuMemFree(outBuf);
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Host<->device transfer bandwidth (CUDA)
-// ---------------------------------------------------------------------------
-
-int CudaPeak::runTransferBandwidth(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  const uint64_t bytes = cfg.transferBWMaxSize ? cfg.transferBWMaxSize : (1ull << 27);
-  unsigned int forced = forceIters ? specifiedIters : 0;
-
-  auto test = currentDeviceScope->beginTest(
-    {"transfer_bandwidth", "Transfer bandwidth (GBPS)", "gbps"});
-
-  CUdeviceptr dBuf = 0;
-  if (cuMemAlloc(&dBuf, bytes) != CUDA_SUCCESS)
-  {
-    test.skip("h2d_pinned", ResultStatus::Error, "Failed to allocate device buffer");
-    test.skip("d2h_pinned", ResultStatus::Error, "Failed to allocate device buffer");
-    return -1;
-  }
-  void *hPinned = nullptr;
-  if (cuMemAllocHost(&hPinned, bytes) != CUDA_SUCCESS)
-  {
-    cuMemFree(dBuf);
-    test.skip("h2d_pinned", ResultStatus::Error, "Failed to allocate pinned host buffer");
-    test.skip("d2h_pinned", ResultStatus::Error, "Failed to allocate pinned host buffer");
-    return -1;
-  }
-
-  auto timeXfer = [&](bool h2d) -> float
-  {
-    CUevent s, e;
-    cuEventCreate(&s, CU_EVENT_DEFAULT);
-    cuEventCreate(&e, CU_EVENT_DEFAULT);
-
-    auto runBatch = [&](unsigned int n) -> float {
-      cuEventRecord(s, dev.stream);
-      for (unsigned i = 0; i < n; i++)
-      {
-        if (h2d)
-          cuMemcpyHtoDAsync(dBuf, hPinned, bytes, dev.stream);
-        else
-          cuMemcpyDtoHAsync(hPinned, dBuf, bytes, dev.stream);
-      }
-      cuEventRecord(e, dev.stream);
-      cuEventSynchronize(e);
-      float ms = 0;
-      cuEventElapsedTime(&ms, s, e);
-      return ms * 1000.0f; // total us
-    };
-
-    // Phase 1: untimed warmup.
-    for (unsigned w = 0; w < warmupCount; w++)
-    {
-      if (h2d)
-        cuMemcpyHtoDAsync(dBuf, hPinned, bytes, dev.stream);
-      else
-        cuMemcpyDtoHAsync(hPinned, dBuf, bytes, dev.stream);
-      cuStreamSynchronize(dev.stream);
-    }
-
-    // Phase 2: timed probe -> per-iter time -> calibrated iters.
-    unsigned int probeIters = 1;
-    float probeUs = runBatch(probeIters);
-    double per_iter_us = (double)probeUs / (double)probeIters;
-    unsigned int iters = pickIters(per_iter_us, cfg.targetTimeUs, forced);
-
-    // Phase 3: real timed run.
-    float totalUs = runBatch(iters);
-
-    cuEventDestroy(s);
-    cuEventDestroy(e);
-    return totalUs / iters; // microseconds per transfer
-  };
-
-  float usH2D = timeXfer(true);
-  float gbpsH2D = (float)bytes / usH2D / 1e3f;
-  test.emit("h2d_pinned", gbpsH2D);
-
-  float usD2H = timeXfer(false);
-  float gbpsD2H = (float)bytes / usD2H / 1e3f;
-  test.emit("d2h_pinned", gbpsD2H);
-
-  cuMemFreeHost(hPinned);
-  cuMemFree(dBuf);
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Kernel launch latency (CUDA)
-// ---------------------------------------------------------------------------
-
-int CudaPeak::runKernelLatency(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  unsigned int iters = forceIters ? specifiedIters
-                                  : (cfg.kernelLatencyIters ? cfg.kernelLatencyIters : 1000);
-
-  auto test = currentDeviceScope->beginTest(
-    {"kernel_launch_latency", "Kernel launch latency (us)", "us"});
-
-  CUfunction fn;
-  if (!dev.getKernel(cuda_kernels::kernel_latency_src,
-                     cuda_kernels::kernel_latency_name,
-                     "kernel_latency_noop", fn))
-  {
-    test.skip("dispatch", ResultStatus::Error, "Kernel compile failed");
-    test.skip("roundtrip", ResultStatus::Error, "Kernel compile failed");
-    return -1;
-  }
-
-  void *args[1] = {nullptr};
-
-  // CUDA's driver API has no primitive equivalent to OpenCL's QUEUED -> START
-  // profiling info or VK_EXT_calibrated_timestamps -- cuEventRecord captures
-  // GPU stream-time, with no portable way to project a host timestamp into
-  // the same domain.  So we report only the round-trip metric here, leaving
-  // dispatch latency as "not measurable" rather than a misleading estimate.
-
-  bool submitFailed = false;
-
-  // Warmup
-  for (unsigned int w = 0; w < warmupCount; w++)
-  {
-    CUresult lr = cuLaunchKernel(fn, 1, 1, 1, 1, 1, 1, 0, dev.stream, args, nullptr);
-    CUresult sr = cuStreamSynchronize(dev.stream);
-    if (lr != CUDA_SUCCESS || sr != CUDA_SUCCESS) { submitFailed = true; break; }
-  }
-
-  double totalRoundtripUs = 0;
-  if (!submitFailed)
-  {
-    for (unsigned int i = 0; i < iters; i++)
-    {
-      auto t0 = std::chrono::high_resolution_clock::now();
-      CUresult lr = cuLaunchKernel(fn, 1, 1, 1, 1, 1, 1, 0, dev.stream, args, nullptr);
-      CUresult sr = cuStreamSynchronize(dev.stream);
-      auto t1 = std::chrono::high_resolution_clock::now();
-      if (lr != CUDA_SUCCESS || sr != CUDA_SUCCESS) { submitFailed = true; break; }
-      totalRoundtripUs += (double)std::chrono::duration_cast<std::chrono::nanoseconds>(t1 - t0).count() / 1000.0;
-    }
-  }
-
-  test.skip("dispatch", ResultStatus::Unsupported,
-            "Not measurable via CUDA driver API");
-  if (submitFailed)
-  {
-    test.skip("roundtrip", ResultStatus::Error,
-              "cuLaunchKernel/cuStreamSynchronize failed");
-  }
-  else
-  {
-    float roundtripUs = (float)(totalRoundtripUs / iters);
-    test.emit("roundtrip", roundtripUs);
-  }
-
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Local memory bandwidth (CUDA -- __shared__ memory)
-// ---------------------------------------------------------------------------
-
-int CudaPeak::runLocalBandwidth(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  auto test = currentDeviceScope->beginTest(
-    {"local_memory_bandwidth", "Local memory bandwidth (GBPS)", "gbps"});
-
-  const uint32_t blockSize = 256;
-  uint64_t globalThreads = targetGlobalThreads((uint32_t)dev.info.numSMs);
-  uint32_t numBlocks = (uint32_t)(globalThreads / blockSize);
-
-  CUdeviceptr outBuf = 0;
-  if (cuMemAlloc(&outBuf, globalThreads * sizeof(float)) != CUDA_SUCCESS)
-  {
-    test.skip("float", ResultStatus::Error, "Buffer alloc failed");
-    test.skip("float2", ResultStatus::Error, "Buffer alloc failed");
-    test.skip("float4", ResultStatus::Error, "Buffer alloc failed");
-    return -1;
-  }
-
-  struct V
-  {
-    const char *label;
-    const char *kname;
-    uint32_t width;
-  };
-  const V vs[] = {
-      {"float  ", "local_bandwidth_v1", 1},
-      {"float2 ", "local_bandwidth_v2", 2},
-      {"float4 ", "local_bandwidth_v4", 4},
-  };
-  for (const auto &v : vs)
-  {
-    CUfunction fn;
-    if (!dev.getKernel(cuda_kernels::local_bandwidth_src,
-                       cuda_kernels::local_bandwidth_name, v.kname, fn))
-    {
-      std::string key(v.label);
-      while (!key.empty() && key.back() == ' ')
-        key.pop_back();
-      test.skip(key, ResultStatus::Error, "Kernel compile failed");
-      continue;
-    }
-    void *args[1] = {&outBuf};
-    float us = runKernel(dev, fn, numBlocks, blockSize, args,
-                         cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-    uint64_t bytes = (uint64_t)LMEM_REPS * 2 * v.width * sizeof(float) * globalThreads;
-    float gbps = (float)bytes / us / 1e3f;
-    std::string key(v.label);
-    while (!key.empty() && key.back() == ' ')
-      key.pop_back();
-    test.emit(key, gbps);
-  }
-
-  cuMemFree(outBuf);
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Image (texture) bandwidth (CUDA -- cudaTextureObject_t via driver API)
-// ---------------------------------------------------------------------------
-
-int CudaPeak::runImageBandwidth(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  auto test = currentDeviceScope->beginTest(
-    {"image_memory_bandwidth", "Image memory bandwidth (GBPS)", "gbps"});
-
-  const int imgW = 4096, imgH = 4096;
-  const uint32_t blockSize = 256;
-  uint64_t globalThreads = targetGlobalThreads((uint32_t)dev.info.numSMs);
-  uint32_t numBlocks = (uint32_t)(globalThreads / blockSize);
-
-  // Create CUarray (RGBA float).
-  CUDA_ARRAY_DESCRIPTOR adesc = {};
-  adesc.Width = imgW;
-  adesc.Height = imgH;
-  adesc.Format = CU_AD_FORMAT_FLOAT;
-  adesc.NumChannels = 4;
-  CUarray arr;
-  if (cuArrayCreate(&arr, &adesc) != CUDA_SUCCESS)
-  {
-    test.skip("float4", ResultStatus::Error, "Image array create failed");
-    return -1;
-  }
-  // Contents undefined is fine for a bandwidth measurement -- the cache
-  // lines still get fetched.
-
-  CUDA_RESOURCE_DESC rd = {};
-  rd.resType = CU_RESOURCE_TYPE_ARRAY;
-  rd.res.array.hArray = arr;
-  CUDA_TEXTURE_DESC td = {};
-  td.addressMode[0] = CU_TR_ADDRESS_MODE_CLAMP;
-  td.addressMode[1] = CU_TR_ADDRESS_MODE_CLAMP;
-  td.filterMode = CU_TR_FILTER_MODE_POINT;
-  td.flags = CU_TRSF_READ_AS_INTEGER; // we want raw float bits, no normalization
-  CUtexObject tex = 0;
-  if (cuTexObjectCreate(&tex, &rd, &td, nullptr) != CUDA_SUCCESS)
-  {
-    cuArrayDestroy(arr);
-    test.skip("float4", ResultStatus::Error, "Texture object create failed");
-    return -1;
-  }
-
-  CUdeviceptr outBuf = 0;
-  cuMemAlloc(&outBuf, globalThreads * sizeof(float));
-
-  CUfunction fn;
-  if (!dev.getKernel(cuda_kernels::image_bandwidth_src,
-                     cuda_kernels::image_bandwidth_name,
-                     "image_bandwidth", fn))
-  {
-    test.skip("float4", ResultStatus::Error, "Kernel compile failed");
-    cuTexObjectDestroy(tex);
-    cuArrayDestroy(arr);
-    cuMemFree(outBuf);
-    return -1;
-  }
-
-  int w = imgW, h = imgH;
-  void *args[4] = {&tex, &outBuf, &w, &h};
-  float us = runKernel(dev, fn, numBlocks, blockSize, args,
-                       cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-  uint64_t bytes = (uint64_t)IMAGE_FETCH_PER_WI * 4 * sizeof(float) * globalThreads;
-  float gbps = (float)bytes / us / 1e3f;
-  test.emit("float4", gbps);
-
-  cuTexObjectDestroy(tex);
-  cuArrayDestroy(arr);
-  cuMemFree(outBuf);
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Atomic throughput (CUDA -- global + local atomics)
-// ---------------------------------------------------------------------------
-
-int CudaPeak::runAtomicThroughput(CudaDevice &dev, benchmark_config_t &cfg)
-{
-  auto test = currentDeviceScope->beginTest(
-    {"atomic_throughput", "Atomic throughput (GOPS)", "gops"});
-
-  const uint32_t blockSize = 256;
-  uint64_t globalThreads = targetGlobalThreads((uint32_t)dev.info.numSMs);
-  uint32_t numBlocks = (uint32_t)(globalThreads / blockSize);
-
-  // Global: per-thread counter (128 MB).
-  {
-    CUdeviceptr buf = 0;
-    if (cuMemAlloc(&buf, globalThreads * sizeof(int)) == CUDA_SUCCESS)
-    {
-      cuMemsetD32(buf, 0, globalThreads);
-      CUfunction fn;
-      if (dev.getKernel(cuda_kernels::atomic_throughput_src,
-                        cuda_kernels::atomic_throughput_name,
-                        "atomic_throughput_global", fn))
-      {
-        void *args[1] = {&buf};
-        float us = runKernel(dev, fn, numBlocks, blockSize, args,
-                             cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-        float gops = ((float)globalThreads * (float)ATOMIC_REPS) / us / 1e3f;
-        test.emit("int_global", gops);
-      }
-      else
-      {
-        test.skip("int_global", ResultStatus::Error, "Kernel compile failed");
-      }
-      cuMemFree(buf);
-    }
-    else
-    {
-      test.skip("int_global", ResultStatus::Error, "Buffer alloc failed");
-    }
-  }
-
-  // Local: one counter per block.
-  {
-    CUdeviceptr buf = 0;
-    if (cuMemAlloc(&buf, (uint64_t)numBlocks * sizeof(int)) == CUDA_SUCCESS)
-    {
-      CUfunction fn;
-      if (dev.getKernel(cuda_kernels::atomic_throughput_src,
-                        cuda_kernels::atomic_throughput_name,
-                        "atomic_throughput_local", fn))
-      {
-        void *args[1] = {&buf};
-        float us = runKernel(dev, fn, numBlocks, blockSize, args,
-                             cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-        float gops = ((float)globalThreads * (float)ATOMIC_REPS) / us / 1e3f;
-        test.emit("int_local", gops);
-      }
-      else
-      {
-        test.skip("int_local", ResultStatus::Error, "Kernel compile failed");
-      }
-      cuMemFree(buf);
-    }
-    else
-    {
-      test.skip("int_local", ResultStatus::Error, "Buffer alloc failed");
-    }
-  }
-
-  return 0;
-}
 
 // Free-function enumeration used by --list-devices and the Android JNI surface.
 // Uses the static driver API directly — no CudaPeak instance required.
