+++ src/vulkan/vk_peak.cpp
@@ -14,22 +14,6 @@
 #include <time.h>
 #endif
 
-static uint64_t targetVulkanGlobalThreads(const vk_device_info_t &info)
-{
-  if (info.numCUs > 0)
-    return targetGlobalThreads(info.numCUs);
-
-  // Mobile/integrated Vulkan drivers often do not expose a vendor CU-count
-  // property.  The desktop 32M fallback can make the calibration probe itself
-  // a multi-second dispatch on those GPUs, so start smaller and let timed
-  // calibration batch more dispatches when the kernel is fast enough.
-  if (info.vkDeviceType == VK_PHYSICAL_DEVICE_TYPE_INTEGRATED_GPU ||
-      info.vkDeviceType == VK_PHYSICAL_DEVICE_TYPE_CPU)
-    return 2ULL << 20;
-
-  return targetGlobalThreads(0);
-}
-
 // ---------------------------------------------------------------------------
 // VulkanDevice
 // ---------------------------------------------------------------------------
@@ -1159,1601 +1143,11 @@ int vkPeak::runComputeKernel(VulkanDevice &dev, benchmark_config_t &cfg,
 }
 
 // ---------------------------------------------------------------------------
-// Concrete compute benchmarks.  Each is a thin wrapper that fills in a
-// vk_compute_desc_t and delegates to runComputeKernel above.
+// Benchmark methods live in separate category files:
+//   compute_float.cpp    compute_int.cpp    coopmat.cpp
+//   global_bandwidth.cpp local_bandwidth.cpp image_bandwidth.cpp
+//   transfer_bandwidth.cpp atomic_throughput.cpp kernel_latency.cpp
 // ---------------------------------------------------------------------------
-
-int vkPeak::runComputeSP(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  static const vk_compute_variant_t variants[] = {
-    { "float",  vk_shaders::compute_sp_v1, vk_shaders::compute_sp_v1_size },
-#ifdef VK_HAS_COMPUTE_SP_V2
-    { "float2", vk_shaders::compute_sp_v2, vk_shaders::compute_sp_v2_size },
-#endif
-#ifdef VK_HAS_COMPUTE_SP_V4
-    { "float4", vk_shaders::compute_sp_v4, vk_shaders::compute_sp_v4_size },
-#endif
-  };
-  float A = 1.3f;
-  vk_compute_desc_t d = {};
-  d.title       = "Single-precision compute (GFLOPS)";
-  d.resultTag   = "single_precision_compute";
-  d.unit        = "gflops";
-  d.variants    = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI   = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize    = sizeof(float);
-  d.pushData    = &A;
-  d.pushSize    = sizeof(A);
-  return runComputeKernel(dev, cfg, d);
-}
-
-#ifdef VK_HAS_COMPUTE_HP_V1
-int vkPeak::runComputeHP(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  static const vk_compute_variant_t variants[] = {
-    { "half",   vk_shaders::compute_hp_v1, vk_shaders::compute_hp_v1_size },
-#ifdef VK_HAS_COMPUTE_HP_V2
-    { "half2",  vk_shaders::compute_hp_v2, vk_shaders::compute_hp_v2_size },
-#endif
-#ifdef VK_HAS_COMPUTE_HP_V4
-    { "half4",  vk_shaders::compute_hp_v4, vk_shaders::compute_hp_v4_size },
-#endif
-  };
-  float A = 1.3f;
-  vk_compute_desc_t d = {};
-  d.title       = "Half-precision compute fp16xfp16+fp16 (GFLOPS)";
-  d.resultTag   = "half_precision_compute";
-  d.unit        = "gflops";
-  d.variants    = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI   = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize    = sizeof(float);
-  d.pushData    = &A;
-  d.pushSize    = sizeof(A);
-  d.skip        = !dev.info.float16Supported;
-  d.skipMsg     = "shaderFloat16 not supported! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-#endif
-
-#ifdef VK_HAS_COMPUTE_DP_V1
-int vkPeak::runComputeDP(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  static const vk_compute_variant_t variants[] = {
-    { "double",  vk_shaders::compute_dp_v1, vk_shaders::compute_dp_v1_size },
-#ifdef VK_HAS_COMPUTE_DP_V2
-    { "double2", vk_shaders::compute_dp_v2, vk_shaders::compute_dp_v2_size },
-#endif
-#ifdef VK_HAS_COMPUTE_DP_V4
-    { "double4", vk_shaders::compute_dp_v4, vk_shaders::compute_dp_v4_size },
-#endif
-  };
-  double A = 1.3;
-  vk_compute_desc_t d = {};
-  d.title       = "Double-precision compute (GFLOPS)";
-  d.resultTag   = "double_precision_compute";
-  d.unit        = "gflops";
-  d.variants    = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI   = COMPUTE_DP_WORK_PER_WI;
-  d.elemSize    = sizeof(double);
-  d.pushData    = &A;
-  d.pushSize    = sizeof(A);
-  d.skip        = !dev.info.float64Supported;
-  d.skipMsg     = "shaderFloat64 not supported! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-#endif
-
-#ifdef VK_HAS_COMPUTE_INT32_V1
-int vkPeak::runComputeInt32(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  static const vk_compute_variant_t variants[] = {
-    { "int",   vk_shaders::compute_int32_v1, vk_shaders::compute_int32_v1_size },
-#ifdef VK_HAS_COMPUTE_INT32_V2
-    { "int2",  vk_shaders::compute_int32_v2, vk_shaders::compute_int32_v2_size },
-#endif
-#ifdef VK_HAS_COMPUTE_INT32_V4
-    { "int4",  vk_shaders::compute_int32_v4, vk_shaders::compute_int32_v4_size },
-#endif
-  };
-  int32_t A = 4;
-  vk_compute_desc_t d = {};
-  d.title       = "Integer compute int32 (GOPS)";
-  d.resultTag   = "integer_compute";
-  d.unit        = "gops";
-  d.variants    = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI   = COMPUTE_INT_WORK_PER_WI;
-  d.elemSize    = sizeof(int32_t);
-  d.pushData    = &A;
-  d.pushSize    = sizeof(A);
-  return runComputeKernel(dev, cfg, d);
-}
-#endif
-
-#ifdef VK_HAS_COMPUTE_MP_V1
-int vkPeak::runComputeMP(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  // v1 = scalar fp16 (baseline; no HFMA2 packing).
-  // v2 = f16vec2  (unlocks NVIDIA HFMA2 at 2x FP32 rate on shader cores).
-  // v4 = f16vec4  (wider packing; informs AMD/Intel where issue rate
-  //                exceeds two lanes per slot).
-  static const vk_compute_variant_t variants[] = {
-    { "mp",  vk_shaders::compute_mp_v1, vk_shaders::compute_mp_v1_size },
-#ifdef VK_HAS_COMPUTE_MP_V2
-    { "mp2", vk_shaders::compute_mp_v2, vk_shaders::compute_mp_v2_size },
-#endif
-#ifdef VK_HAS_COMPUTE_MP_V4
-    { "mp4", vk_shaders::compute_mp_v4, vk_shaders::compute_mp_v4_size },
-#endif
-  };
-  float A = 1.3f;
-  vk_compute_desc_t d = {};
-  d.title       = "Mixed-precision compute fp16xfp16+fp32 (GFLOPS)";
-  d.resultTag      = "mixed_precision_compute";
-  d.unit        = "gflops";
-  d.variants    = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI   = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize    = sizeof(float);
-  d.pushData    = &A;
-  d.pushSize    = sizeof(A);
-  d.skip        = !dev.info.float16Supported;
-  d.skipMsg     = "shaderFloat16 not supported! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-#endif
-
-#ifdef VK_HAS_COMPUTE_INT4_PACKED_V1
-int vkPeak::runComputeInt4Packed(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  int32_t A = 3;
-  vk_compute_desc_t d = {};
-  d.title           = "Packed INT4 compute (emulated) (GOPS)";
-  d.resultTag          = "int4_packed_compute";
-  d.metricLabel     = "int4_packed";
-  d.unit            = "gops";
-  d.spirv           = vk_shaders::compute_int4_packed_v1;
-  d.spirvSize       = vk_shaders::compute_int4_packed_v1_size;
-  d.workPerWI       = COMPUTE_INT4_PACKED_WORK_PER_WI;
-  d.elemSize        = sizeof(int32_t);
-  d.pushData        = &A;
-  d.pushSize        = sizeof(A);
-  return runComputeKernel(dev, cfg, d);
-}
-#endif
-
-#ifdef VK_HAS_COMPUTE_INT8_DP_V1
-int vkPeak::runComputeInt8DP(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  // v1 = single dp4a chain (serial through REPACK; dep-stall bound).
-  // v2 = two parallel dp4a chains (double the independent work for the
-  //       instruction issue queue to pipeline).
-  // v4 = four parallel chains (enough to saturate dp4a issue rate on
-  //       NVIDIA Turing+ / AMD RDNA2+ / Intel Xe+).
-  static const vk_compute_variant_t variants[] = {
-    { "int8_dp",  vk_shaders::compute_int8_dp_v1, vk_shaders::compute_int8_dp_v1_size },
-#ifdef VK_HAS_COMPUTE_INT8_DP_V2
-    { "int8_dp2", vk_shaders::compute_int8_dp_v2, vk_shaders::compute_int8_dp_v2_size },
-#endif
-#ifdef VK_HAS_COMPUTE_INT8_DP_V4
-    { "int8_dp4", vk_shaders::compute_int8_dp_v4, vk_shaders::compute_int8_dp_v4_size },
-#endif
-  };
-  int32_t A = 4;
-  vk_compute_desc_t d = {};
-  d.title       = "INT8 dot-product compute (GOPS)";
-  d.resultTag      = "integer_compute_int8_dp";
-  d.unit        = "gops";
-  d.variants    = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI   = COMPUTE_INT8_DP_WORK_PER_WI;
-  d.elemSize    = sizeof(int32_t);
-  d.pushData    = &A;
-  d.pushSize    = sizeof(A);
-  d.skip        = !dev.info.int8DotProductSupported;
-  d.skipMsg     = "VK_KHR_shader_integer_dot_product / shaderInt8 not supported! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-#endif
-
-#ifdef VK_HAS_COMPUTE_BF16_V1
-int vkPeak::runComputeBF16(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  // v1 / v2 / v4: same packing story as MP.  NVIDIA shader-core BF16
-  // peaks at bf16vec2 via BMMA2-style packed multiply.
-  static const vk_compute_variant_t variants[] = {
-    { "bf16",  vk_shaders::compute_bf16_v1, vk_shaders::compute_bf16_v1_size },
-#ifdef VK_HAS_COMPUTE_BF16_V2
-    { "bf16_2", vk_shaders::compute_bf16_v2, vk_shaders::compute_bf16_v2_size },
-#endif
-#ifdef VK_HAS_COMPUTE_BF16_V4
-    { "bf16_4", vk_shaders::compute_bf16_v4, vk_shaders::compute_bf16_v4_size },
-#endif
-  };
-  float A = 1.3f;
-  vk_compute_desc_t d = {};
-  d.title       = "BF16 compute bf16xbf16+fp32 (GFLOPS)";
-  d.resultTag      = "bfloat16_compute";
-  d.unit        = "gflops";
-  d.variants    = variants;
-  d.numVariants = sizeof(variants) / sizeof(variants[0]);
-  d.workPerWI   = COMPUTE_FP_WORK_PER_WI;
-  d.elemSize    = sizeof(float);
-  d.pushData    = &A;
-  d.pushSize    = sizeof(A);
-  d.skip        = !dev.info.bfloat16Supported;
-  d.skipMsg     = "VK_KHR_shader_bfloat16 / shaderBFloat16Type not supported! Skipped";
-  return runComputeKernel(dev, cfg, d);
-}
-#endif
-
-// ---------------------------------------------------------------------------
-// Cooperative matrix (tensor-core) umbrella.
-//
-// Runs every dtype combination the driver advertises at the canonical
-// 16x16x16 subgroup-scope tile (queried once in VulkanDevice::init).  Each
-// dtype shares the same scaffolding via runComputeKernel -- only the shader,
-// buffer element type, and label strings differ.  Adding FP8 / INT4 in
-// Phase 2 reduces to: compile a coopmat_fp8.comp, query the matching
-// component-type enums, and add one more entry here.
-// ---------------------------------------------------------------------------
-
-int vkPeak::runCoopMatrix(VulkanDevice &dev, benchmark_config_t &cfg, bool intPart)
-{
-  // Coopmat shape constants: shaders hard-code 16x16x16 with 256 iters and
-  // local_size_x=32 (one subgroup per work-group).  See COOPMAT_WORK_PER_WI.
-  const uint32_t coopWGSize  = 32;
-  const uint32_t coopOutElems = 16 * 16;  // M*N tile written per WG
-  const uint32_t coopWork    = COOPMAT_WORK_PER_WI;
-
-  if (!intPart) {
-#ifdef VK_HAS_COOPMAT_FP32
-    {
-      float A = 1.3f;
-      vk_compute_desc_t d = {};
-      d.title          = "Cooperative-matrix fp32xfp32+fp32 16x16x16 (TFLOPS)";
-      d.resultTag      = "coopmat_fp32";
-      d.metricLabel    = "coopmat_fp32";
-      d.unit           = "tflops";
-      d.unitDivider    = 1e12;
-      d.spirv          = vk_shaders::coopmat_fp32;
-      d.spirvSize      = vk_shaders::coopmat_fp32_size;
-      d.workPerWI      = coopWork;
-      d.elemSize       = sizeof(float);
-      d.wgSize         = coopWGSize;
-      d.outElemsPerWG  = coopOutElems;
-      d.pushData       = &A;
-      d.pushSize       = sizeof(A);
-      d.skip           = !dev.info.coopmatFP32Supported;
-      d.skipMsg        = "No 16x16x16 fp32xfp32+fp32 coopmat property! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-#endif
-#ifdef VK_HAS_COOPMAT_FP16
-    {
-      float A = 1.3f;
-      vk_compute_desc_t d = {};
-      d.title          = "Cooperative-matrix fp16xfp16+fp32 16x16x16 (TFLOPS)";
-      d.resultTag      = "coopmat_fp16";
-      d.metricLabel    = "coopmat_fp16";
-      d.unit           = "tflops";
-      d.unitDivider    = 1e12;
-      d.spirv          = vk_shaders::coopmat_fp16;
-      d.spirvSize      = vk_shaders::coopmat_fp16_size;
-      d.workPerWI      = coopWork;
-      d.elemSize       = sizeof(float);
-      d.wgSize         = coopWGSize;
-      d.outElemsPerWG  = coopOutElems;
-      d.pushData       = &A;
-      d.pushSize       = sizeof(A);
-      d.skip           = !dev.info.coopmatFP16Supported;
-      d.skipMsg        = "No 16x16x16 fp16xfp16+fp32 coopmat property! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-#endif
-#ifdef VK_HAS_COOPMAT_BF16
-    {
-      float A = 1.3f;
-      vk_compute_desc_t d = {};
-      d.title          = "Cooperative-matrix bf16xbf16+fp32 16x16x16 (TFLOPS)";
-      d.resultTag      = "coopmat_bf16";
-      d.metricLabel    = "coopmat_bf16";
-      d.unit           = "tflops";
-      d.unitDivider    = 1e12;
-      d.spirv          = vk_shaders::coopmat_bf16;
-      d.spirvSize      = vk_shaders::coopmat_bf16_size;
-      d.workPerWI      = coopWork;
-      d.elemSize       = sizeof(float);
-      d.wgSize         = coopWGSize;
-      d.outElemsPerWG  = coopOutElems;
-      d.pushData       = &A;
-      d.pushSize       = sizeof(A);
-      d.skip           = !dev.info.coopmatBF16Supported;
-      d.skipMsg        = "No 16x16x16 bf16xbf16+fp32 coopmat property! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-#endif
-#ifdef VK_HAS_COOPMAT_FP8_E4M3
-    {
-      float A = 1.3f;
-      vk_compute_desc_t d = {};
-      d.title          = "Cooperative-matrix fp8(E4M3)xfp8(E4M3)+fp32 16x16x16 (TFLOPS)";
-      d.resultTag      = "coopmat_fp8_e4m3";
-      d.metricLabel    = "coopmat_fp8_e4m3";
-      d.unit           = "tflops";
-      d.unitDivider    = 1e12;
-      d.spirv          = vk_shaders::coopmat_fp8_e4m3;
-      d.spirvSize      = vk_shaders::coopmat_fp8_e4m3_size;
-      d.workPerWI      = coopWork;
-      d.elemSize       = sizeof(float);
-      d.wgSize         = coopWGSize;
-      d.outElemsPerWG  = coopOutElems;
-      d.pushData       = &A;
-      d.pushSize       = sizeof(A);
-      d.skip           = !(dev.info.fp8Supported && dev.info.coopmatFP8E4M3Supported);
-      d.skipMsg        = "No fp8-E4M3 coopmat support (VK_EXT_shader_float8 or property)! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-#endif
-#ifdef VK_HAS_COOPMAT_FP8_E5M2
-    {
-      float A = 1.3f;
-      vk_compute_desc_t d = {};
-      d.title          = "Cooperative-matrix fp8(E5M2)xfp8(E5M2)+fp32 16x16x16 (TFLOPS)";
-      d.resultTag      = "coopmat_fp8_e5m2";
-      d.metricLabel    = "coopmat_fp8_e5m2";
-      d.unit           = "tflops";
-      d.unitDivider    = 1e12;
-      d.spirv          = vk_shaders::coopmat_fp8_e5m2;
-      d.spirvSize      = vk_shaders::coopmat_fp8_e5m2_size;
-      d.workPerWI      = coopWork;
-      d.elemSize       = sizeof(float);
-      d.wgSize         = coopWGSize;
-      d.outElemsPerWG  = coopOutElems;
-      d.pushData       = &A;
-      d.pushSize       = sizeof(A);
-      d.skip           = !(dev.info.fp8Supported && dev.info.coopmatFP8E5M2Supported);
-      d.skipMsg        = "No fp8-E5M2 coopmat support (VK_EXT_shader_float8 or property)! Skipped";
-      runComputeKernel(dev, cfg, d);
-    }
-#endif
-  } // !intPart
-
-#if defined(VK_HAS_COOPMAT_INT8) || defined(VK_HAS_COOPMAT_INT8_K32)
-  if (intPart) {
-    // Select the shader variant matching whichever INT8 tile the driver
-    // advertised.  K=16 is the generic path; NVIDIA tensor cores need K=32.
-    int32_t A = 3;
-    vk_compute_desc_t d = {};
-    d.resultTag         = "coopmat_int8";
-    d.metricLabel    = "coopmat_int8";
-    d.unit           = "tops";
-    d.unitDivider    = 1e12;
-    d.workPerWI      = coopWork;
-    d.elemSize       = sizeof(int32_t);
-    d.wgSize         = coopWGSize;
-    d.outElemsPerWG  = coopOutElems;
-    d.pushData       = &A;
-    d.pushSize       = sizeof(A);
-
-    const char *titleK16 = "Cooperative-matrix int8xint8+int32 16x16x16 (TOPS)";
-    const char *titleK32 = "Cooperative-matrix int8xint8+int32 16x16x32 (TOPS)";
-    bool haveShaderK16 = false, haveShaderK32 = false;
-#ifdef VK_HAS_COOPMAT_INT8
-    haveShaderK16 = true;
-#endif
-#ifdef VK_HAS_COOPMAT_INT8_K32
-    haveShaderK32 = true;
-#endif
-
-    if (dev.info.coopmatINT8K == 16 && haveShaderK16)
-    {
-#ifdef VK_HAS_COOPMAT_INT8
-      d.title          = titleK16;
-      d.spirv          = vk_shaders::coopmat_int8;
-      d.spirvSize      = vk_shaders::coopmat_int8_size;
-#endif
-    }
-    else if (dev.info.coopmatINT8K == 32 && haveShaderK32)
-    {
-#ifdef VK_HAS_COOPMAT_INT8_K32
-      d.title          = titleK32;
-      d.spirv          = vk_shaders::coopmat_int8_k32;
-      d.spirvSize      = vk_shaders::coopmat_int8_k32_size;
-#endif
-    }
-    else
-    {
-      // Driver advertised neither 16x16x16 nor 16x16x32 INT8 -- or the
-      // corresponding shader didn't compile in this build.  Skip with a
-      // label that names the probed tiles so the reason is obvious.
-      d.title          = titleK16;
-      d.skip           = true;
-      d.skipMsg        = "No 16x16x{16,32} int8xint8+int32 coopmat property! Skipped";
-      d.spirv          = nullptr;
-      d.spirvSize      = 0;
-    }
-    runComputeKernel(dev, cfg, d);
-  } // if (intPart)
-#endif
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Global bandwidth benchmark (Vulkan)
-// ---------------------------------------------------------------------------
-
-int vkPeak::runGlobalBandwidth(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  const uint32_t wgSize = 256;
-
-  uint64_t maxItems = dev.info.maxAllocSize / sizeof(float) / 2;
-  uint64_t numItems = (maxItems / (wgSize * FETCH_PER_WI)) * (wgSize * FETCH_PER_WI);
-  if (numItems > cfg.globalBWMaxSize / sizeof(float))
-    numItems = (cfg.globalBWMaxSize / sizeof(float) / (wgSize * FETCH_PER_WI)) * (wgSize * FETCH_PER_WI);
-
-  uint32_t numGroups = (uint32_t)(numItems / FETCH_PER_WI / wgSize);
-  if (numGroups == 0) numGroups = 1;
-
-  logger::TestSpec testSpec;
-  testSpec.tag = "global_memory_bandwidth";
-  testSpec.display = "Global memory bandwidth";
-  testSpec.unit = "gbps";
-  auto test = currentDeviceScope->beginTest(testSpec);
-
-  // Create input + output buffers
-  VkBuffer inputBuf, outputBuf;
-  VkDeviceMemory inputMem, outputMem;
-
-  if (!dev.createBuffer(numItems * sizeof(float),
-                        VK_BUFFER_USAGE_STORAGE_BUFFER_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-                        VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT,
-                        inputBuf, inputMem) ||
-      !dev.createBuffer(numItems * sizeof(float),
-                        VK_BUFFER_USAGE_STORAGE_BUFFER_BIT,
-                        VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT,
-                        outputBuf, outputMem))
-  {
-    log->note("Failed to allocate buffers\n");
-    return -1;
-  }
-
-  // Descriptor set layout (2 bindings: input + output)
-  VkDescriptorSetLayoutBinding bindings[2] = {};
-  bindings[0].binding = 0;
-  bindings[0].descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-  bindings[0].descriptorCount = 1;
-  bindings[0].stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
-  bindings[1].binding = 1;
-  bindings[1].descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-  bindings[1].descriptorCount = 1;
-  bindings[1].stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
-
-  VkDescriptorSetLayoutCreateInfo dsLayoutCI = {};
-  dsLayoutCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-  dsLayoutCI.bindingCount = 2;
-  dsLayoutCI.pBindings = bindings;
-
-  VkDescriptorSetLayout dsLayout;
-  vkCreateDescriptorSetLayout(dev.device, &dsLayoutCI, nullptr, &dsLayout);
-
-  VkPipelineLayoutCreateInfo pipeLayoutCI = {};
-  pipeLayoutCI.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-  pipeLayoutCI.setLayoutCount = 1;
-  pipeLayoutCI.pSetLayouts = &dsLayout;
-
-  VkPipelineLayout pipeLayout;
-  vkCreatePipelineLayout(dev.device, &pipeLayoutCI, nullptr, &pipeLayout);
-
-  struct GBVar { const char *label; const uint32_t *spv; size_t sz; uint32_t width; };
-  const GBVar variants[] = {
-    {"float ",  vk_shaders::global_bandwidth_v1, vk_shaders::global_bandwidth_v1_size, 1},
-#ifdef VK_HAS_GLOBAL_BANDWIDTH_V2
-    {"float2",  vk_shaders::global_bandwidth_v2, vk_shaders::global_bandwidth_v2_size, 2},
-#endif
-#ifdef VK_HAS_GLOBAL_BANDWIDTH_V4
-    {"float4",  vk_shaders::global_bandwidth_v4, vk_shaders::global_bandwidth_v4_size, 4},
-#endif
-  };
-
-  // Descriptor pool + set
-  VkDescriptorPoolSize poolSize = {};
-  poolSize.type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-  poolSize.descriptorCount = 2;
-
-  VkDescriptorPoolCreateInfo dpCI = {};
-  dpCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-  dpCI.maxSets = 1;
-  dpCI.poolSizeCount = 1;
-  dpCI.pPoolSizes = &poolSize;
-
-  VkDescriptorPool descPool;
-  vkCreateDescriptorPool(dev.device, &dpCI, nullptr, &descPool);
-
-  VkDescriptorSetAllocateInfo dsAI = {};
-  dsAI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-  dsAI.descriptorPool = descPool;
-  dsAI.descriptorSetCount = 1;
-  dsAI.pSetLayouts = &dsLayout;
-
-  VkDescriptorSet descSet;
-  vkAllocateDescriptorSets(dev.device, &dsAI, &descSet);
-
-  VkDescriptorBufferInfo bufInfos[2] = {};
-  bufInfos[0].buffer = inputBuf;
-  bufInfos[0].offset = 0;
-  bufInfos[0].range = numItems * sizeof(float);
-  bufInfos[1].buffer = outputBuf;
-  bufInfos[1].offset = 0;
-  bufInfos[1].range = numItems * sizeof(float);
-
-  VkWriteDescriptorSet writes[2] = {};
-  for (int i = 0; i < 2; i++)
-  {
-    writes[i].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-    writes[i].dstSet = descSet;
-    writes[i].dstBinding = i;
-    writes[i].descriptorCount = 1;
-    writes[i].descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    writes[i].pBufferInfo = &bufInfos[i];
-  }
-  vkUpdateDescriptorSets(dev.device, 2, writes, 0, nullptr);
-
-  // Run each variant.  Same buffer + descriptor set; only the pipeline
-  // (i.e. the loaded shader) changes per row.
-  for (const auto &v : variants)
-  {
-    VkPipeline pipe;
-    if (!dev.createComputePipeline(v.spv, v.sz, dsLayout, pipeLayout, pipe))
-    {
-      std::string key(v.label);
-      while (!key.empty() && key.back() == ' ') key.pop_back();
-      test.skip(key, ResultStatus::Error, "Pipeline creation failed");
-      continue;
-    }
-    // Match OpenCL: reduce work-group count by vector width so total bytes
-    // touched stays at numItems * sizeof(float) across all variants.
-    uint32_t variantGroups = numGroups / v.width;
-    if (variantGroups == 0) variantGroups = 1;
-    float timed = runKernel(dev, pipe, pipeLayout, descSet, variantGroups,
-                            cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-    std::string key(v.label);
-    while (!key.empty() && key.back() == ' ') key.pop_back();
-    if (timed <= 0.0f)
-    {
-      test.skip(key, ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-      vkDestroyPipeline(dev.device, pipe, nullptr);
-      continue;
-    }
-    float gbps = ((float)numItems * sizeof(float)) / timed / 1e3f;
-    test.emit(key, gbps);
-    vkDestroyPipeline(dev.device, pipe, nullptr);
-  }
-
-  // Cleanup
-  vkDestroyDescriptorPool(dev.device, descPool, nullptr);
-  vkDestroyPipelineLayout(dev.device, pipeLayout, nullptr);
-  vkDestroyDescriptorSetLayout(dev.device, dsLayout, nullptr);
-  vkDestroyBuffer(dev.device, inputBuf, nullptr);
-  vkFreeMemory(dev.device, inputMem, nullptr);
-  vkDestroyBuffer(dev.device, outputBuf, nullptr);
-  vkFreeMemory(dev.device, outputMem, nullptr);
-
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Host<->device transfer bandwidth (Vulkan)
-// ---------------------------------------------------------------------------
-
-int vkPeak::runTransferBandwidth(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  VkPhysicalDeviceProperties props;
-  vkGetPhysicalDeviceProperties(dev.physicalDevice, &props);
-
-  uint32_t queueFamilyCount = 0;
-  vkGetPhysicalDeviceQueueFamilyProperties(dev.physicalDevice, &queueFamilyCount, nullptr);
-  std::vector<VkQueueFamilyProperties> queueFamilies(queueFamilyCount);
-  vkGetPhysicalDeviceQueueFamilyProperties(dev.physicalDevice, &queueFamilyCount, queueFamilies.data());
-
-  const bool canUseTimestamps =
-      dev.info.computeQueueFamily < queueFamilies.size() &&
-      queueFamilies[dev.info.computeQueueFamily].timestampValidBits > 0 &&
-      props.limits.timestampPeriod > 0.0f;
-
-  uint64_t bytes = cfg.transferBWMaxSize ? cfg.transferBWMaxSize : (1ull << 27);
-  if (dev.info.maxAllocSize > 0)
-    bytes = std::min(bytes, dev.info.maxAllocSize);
-  bytes &= ~255ull;
-  if (bytes == 0)
-    bytes = 256;
-
-  unsigned int forced = forceIters ? specifiedIters : 0;
-
-  logger::TestSpec testSpec;
-  testSpec.tag = "transfer_bandwidth";
-  testSpec.display = "Transfer bandwidth";
-  testSpec.unit = "gbps";
-  auto test = currentDeviceScope->beginTest(testSpec);
-
-  VkBuffer hostBuf = VK_NULL_HANDLE;
-  VkBuffer devBuf = VK_NULL_HANDLE;
-  VkDeviceMemory hostMem = VK_NULL_HANDLE;
-  VkDeviceMemory devMem = VK_NULL_HANDLE;
-
-  if (!dev.createBuffer(bytes,
-                        VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-                        VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT,
-                        hostBuf, hostMem) ||
-      !dev.createBuffer(bytes,
-                        VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-                        VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT,
-                        devBuf, devMem))
-  {
-    log->note("Failed to allocate transfer buffers\n");
-    if (hostBuf != VK_NULL_HANDLE) vkDestroyBuffer(dev.device, hostBuf, nullptr);
-    if (hostMem != VK_NULL_HANDLE) vkFreeMemory(dev.device, hostMem, nullptr);
-    if (devBuf != VK_NULL_HANDLE) vkDestroyBuffer(dev.device, devBuf, nullptr);
-    if (devMem != VK_NULL_HANDLE) vkFreeMemory(dev.device, devMem, nullptr);
-    return -1;
-  }
-
-  VkCommandBufferAllocateInfo allocInfo = {};
-  allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-  allocInfo.commandPool = dev.commandPool;
-  allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-  allocInfo.commandBufferCount = 1;
-
-  VkCommandBuffer cmdBuf = VK_NULL_HANDLE;
-  vkAllocateCommandBuffers(dev.device, &allocInfo, &cmdBuf);
-
-  VkQueryPool queryPool = VK_NULL_HANDLE;
-  if (canUseTimestamps)
-  {
-    VkQueryPoolCreateInfo qpCI = {};
-    qpCI.sType = VK_STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO;
-    qpCI.queryType = VK_QUERY_TYPE_TIMESTAMP;
-    qpCI.queryCount = 2;
-    vkCreateQueryPool(dev.device, &qpCI, nullptr, &queryPool);
-  }
-
-  auto runCopy = [&](VkBuffer src, VkBuffer dst) -> float
-  {
-    VkBufferCopy region = {};
-    region.size = bytes;
-
-    auto recordCopy = [&](bool timed) {
-      VkCommandBufferBeginInfo beginInfo = {};
-      beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-      beginInfo.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-      vkBeginCommandBuffer(cmdBuf, &beginInfo);
-      if (timed && queryPool != VK_NULL_HANDLE)
-      {
-        vkCmdResetQueryPool(cmdBuf, queryPool, 0, 2);
-        vkCmdWriteTimestamp(cmdBuf, VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT, queryPool, 0);
-      }
-      vkCmdCopyBuffer(cmdBuf, src, dst, 1, &region);
-      if (timed && queryPool != VK_NULL_HANDLE)
-        vkCmdWriteTimestamp(cmdBuf, VK_PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT, queryPool, 1);
-      vkEndCommandBuffer(cmdBuf);
-    };
-
-    // Run `n` timed copies, returning total us (or -1 on submit failure).
-    auto runIters = [&](unsigned int n) -> float {
-      float totalUs = 0.0f;
-      for (unsigned int i = 0; i < n; i++)
-      {
-        recordCopy(queryPool != VK_NULL_HANDLE);
-        auto start = std::chrono::high_resolution_clock::now();
-        VkResult sr = dev.submitAndWait(cmdBuf);
-        auto end = std::chrono::high_resolution_clock::now();
-        if (sr != VK_SUCCESS)
-        {
-          vkResetCommandBuffer(cmdBuf, 0);
-          return -1.0f;
-        }
-
-        double iterUs = (double)std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-        if (queryPool != VK_NULL_HANDLE)
-        {
-          uint64_t ts[2] = {};
-          VkResult qr = vkGetQueryPoolResults(dev.device, queryPool, 0, 2, sizeof(ts), ts,
-                                              sizeof(uint64_t),
-                                              VK_QUERY_RESULT_64_BIT | VK_QUERY_RESULT_WAIT_BIT);
-          // Strict > (not >=): some drivers (e.g. Adreno+Turnip) return both
-          // timestamps with the same value for buffer-copy ranges instead of
-          // failing the query.  Treat equality as "timer didn't tick" and
-          // fall back to the host wall clock.
-          if (qr == VK_SUCCESS && ts[1] > ts[0])
-            totalUs += (float)((double)(ts[1] - ts[0]) * props.limits.timestampPeriod / 1000.0);
-          else
-            totalUs += (float)iterUs;
-        }
-        else
-        {
-          totalUs += (float)iterUs;
-        }
-
-        vkResetCommandBuffer(cmdBuf, 0);
-      }
-      return totalUs;
-    };
-
-    // Phase 1: untimed warmup.
-    for (unsigned int w = 0; w < warmupCount; w++)
-    {
-      recordCopy(false);
-      dev.submitAndWait(cmdBuf);
-      vkResetCommandBuffer(cmdBuf, 0);
-    }
-
-    // Phase 2: timed probe -> per-iter time -> calibrated iters.
-    unsigned int probeIters = 1;
-    float probeUs = runIters(probeIters);
-    if (probeUs < 0.0f) return -1.0f;
-    if (probeUs <= 0.0f) probeUs = 1.0f;
-    double per_iter_us = (double)probeUs / (double)probeIters;
-    unsigned int iters = pickIters(per_iter_us, cfg.targetTimeUs, forced);
-
-    // Phase 3: real timed run.
-    float totalUs = runIters(iters);
-    if (totalUs < 0.0f) return -1.0f;
-    // A genuine zero reading (no measurable wall-clock time AND no
-    // timestamp delta) is the timer-broken case -- don't paper over it
-    // with FLT_EPSILON, since (bytes / FLT_EPSILON / 1e3) prints 1.13e12
-    // GBPS and looks like a real result.
-    if (totalUs <= 0.0f)
-      return -1.0f;
-    return totalUs / static_cast<float>(iters);
-  };
-
-  auto reportCopy = [&](const char *metric, float us)
-  {
-    if (us <= 0.0f)
-    {
-      test.skip(metric, ResultStatus::Error,
-                 "vkQueueSubmit/WaitIdle failed or timer returned zero");
-      return;
-    }
-    float gbps = (float)bytes / us / 1e3f;
-    test.emit(metric, gbps);
-  };
-
-  reportCopy("h2d", runCopy(hostBuf, devBuf));
-  reportCopy("d2h", runCopy(devBuf, hostBuf));
-
-  if (queryPool != VK_NULL_HANDLE)
-    vkDestroyQueryPool(dev.device, queryPool, nullptr);
-  vkFreeCommandBuffers(dev.device, dev.commandPool, 1, &cmdBuf);
-  vkDestroyBuffer(dev.device, hostBuf, nullptr);
-  vkFreeMemory(dev.device, hostMem, nullptr);
-  vkDestroyBuffer(dev.device, devBuf, nullptr);
-  vkFreeMemory(dev.device, devMem, nullptr);
-
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Local memory bandwidth (Vulkan -- shared memory)
-// ---------------------------------------------------------------------------
-//
-// Same single-output-buffer scaffolding as runComputeKernel; only the bytes-
-// per-WI calculation differs per variant.  Width = 1/2/4/8 floats per slot.
-
-int vkPeak::runLocalBandwidth(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  logger::TestSpec testSpec;
-  testSpec.tag = "local_memory_bandwidth";
-  testSpec.display = "Local memory bandwidth";
-  testSpec.unit = "gbps";
-  auto test = currentDeviceScope->beginTest(testSpec);
-
-  const uint32_t wgSize = 256;
-  uint64_t globalWIs = targetVulkanGlobalThreads(dev.info);
-  uint32_t numGroups = (uint32_t)(globalWIs / wgSize);
-  uint64_t bufferBytes = (uint64_t)globalWIs * sizeof(float);
-
-  VkBuffer outBuf;
-  VkDeviceMemory outMem;
-  if (!dev.createBuffer(bufferBytes, VK_BUFFER_USAGE_STORAGE_BUFFER_BIT,
-                        VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, outBuf, outMem))
-  {
-    log->note("Failed to allocate buffer\n");
-    return -1;
-  }
-
-  // Descriptor set layout: 1 storage buffer.
-  VkDescriptorSetLayoutBinding binding = {};
-  binding.binding = 0;
-  binding.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-  binding.descriptorCount = 1;
-  binding.stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
-
-  VkDescriptorSetLayoutCreateInfo dslCI = {};
-  dslCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-  dslCI.bindingCount = 1; dslCI.pBindings = &binding;
-  VkDescriptorSetLayout dsLayout;
-  vkCreateDescriptorSetLayout(dev.device, &dslCI, nullptr, &dsLayout);
-
-  VkPipelineLayoutCreateInfo plCI = {};
-  plCI.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-  plCI.setLayoutCount = 1; plCI.pSetLayouts = &dsLayout;
-  VkPipelineLayout pipeLayout;
-  vkCreatePipelineLayout(dev.device, &plCI, nullptr, &pipeLayout);
-
-  VkDescriptorPoolSize ps = {VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1};
-  VkDescriptorPoolCreateInfo dpCI = {};
-  dpCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-  dpCI.maxSets = 1; dpCI.poolSizeCount = 1; dpCI.pPoolSizes = &ps;
-  VkDescriptorPool descPool;
-  vkCreateDescriptorPool(dev.device, &dpCI, nullptr, &descPool);
-
-  VkDescriptorSetAllocateInfo dsAI = {};
-  dsAI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-  dsAI.descriptorPool = descPool; dsAI.descriptorSetCount = 1; dsAI.pSetLayouts = &dsLayout;
-  VkDescriptorSet descSet;
-  vkAllocateDescriptorSets(dev.device, &dsAI, &descSet);
-
-  VkDescriptorBufferInfo bi = {outBuf, 0, bufferBytes};
-  VkWriteDescriptorSet w = {};
-  w.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-  w.dstSet = descSet; w.dstBinding = 0; w.descriptorCount = 1;
-  w.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER; w.pBufferInfo = &bi;
-  vkUpdateDescriptorSets(dev.device, 1, &w, 0, nullptr);
-
-  struct V { const char *label; const uint32_t *spv; size_t sz; uint32_t width; };
-  const V variants[] = {
-    {"float  ", vk_shaders::local_bandwidth_v1, vk_shaders::local_bandwidth_v1_size, 1},
-    {"float2 ", vk_shaders::local_bandwidth_v2, vk_shaders::local_bandwidth_v2_size, 2},
-    {"float4 ", vk_shaders::local_bandwidth_v4, vk_shaders::local_bandwidth_v4_size, 4},
-  };
-  for (const auto &v : variants)
-  {
-    VkPipeline pipe;
-    if (!dev.createComputePipeline(v.spv, v.sz, dsLayout, pipeLayout, pipe))
-    {
-      std::string key(v.label);
-      while (!key.empty() && key.back() == ' ') key.pop_back();
-      test.skip(key, ResultStatus::Error, "Pipeline creation failed");
-      continue;
-    }
-    float us = runKernel(dev, pipe, pipeLayout, descSet, numGroups,
-                         cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-    // Strip padding from the display label for the result metric key.
-    std::string key(v.label);
-    while (!key.empty() && key.back() == ' ') key.pop_back();
-    if (us <= 0.0f)
-    {
-      test.skip(key, ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-      vkDestroyPipeline(dev.device, pipe, nullptr);
-      continue;
-    }
-    // Each rep: 1 write + 1 read per WI = 2 * width * sizeof(float) bytes.
-    uint64_t bytes = (uint64_t)LMEM_REPS * 2 * v.width * sizeof(float) * globalWIs;
-    float gbps = (float)bytes / us / 1e3f;
-    test.emit(key, gbps);
-    vkDestroyPipeline(dev.device, pipe, nullptr);
-  }
-
-  vkDestroyDescriptorPool(dev.device, descPool, nullptr);
-  vkDestroyPipelineLayout(dev.device, pipeLayout, nullptr);
-  vkDestroyDescriptorSetLayout(dev.device, dsLayout, nullptr);
-  vkDestroyBuffer(dev.device, outBuf, nullptr);
-  vkFreeMemory(dev.device, outMem, nullptr);
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Image (texture) bandwidth (Vulkan)
-// ---------------------------------------------------------------------------
-//
-// Combined image-sampler descriptor + storage-buffer output.  Image is
-// VK_FORMAT_R32G32B32A32_SFLOAT, sampled with NEAREST + CLAMP_TO_EDGE.
-
-int vkPeak::runImageBandwidth(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  logger::TestSpec testSpec;
-  testSpec.tag = "image_memory_bandwidth";
-  testSpec.display = "Image memory bandwidth";
-  testSpec.unit = "gbps";
-  auto test = currentDeviceScope->beginTest(testSpec);
-
-  const uint32_t imgW = 4096, imgH = 4096;
-  const uint32_t wgSize = 256;
-  uint64_t globalWIs = targetVulkanGlobalThreads(dev.info);
-  uint32_t numGroups = (uint32_t)(globalWIs / wgSize);
-  uint64_t outBytes  = globalWIs * sizeof(float);
-
-  // Create image (RGBA32F, sampled, transfer-dst so we can clear it).
-  VkImageCreateInfo imgCI = {};
-  imgCI.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
-  imgCI.imageType = VK_IMAGE_TYPE_2D;
-  imgCI.format    = VK_FORMAT_R32G32B32A32_SFLOAT;
-  imgCI.extent    = {imgW, imgH, 1};
-  imgCI.mipLevels = 1; imgCI.arrayLayers = 1;
-  imgCI.samples   = VK_SAMPLE_COUNT_1_BIT;
-  imgCI.tiling    = VK_IMAGE_TILING_OPTIMAL;
-  imgCI.usage     = VK_IMAGE_USAGE_SAMPLED_BIT | VK_IMAGE_USAGE_TRANSFER_DST_BIT;
-  imgCI.sharingMode   = VK_SHARING_MODE_EXCLUSIVE;
-  imgCI.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-
-  VkImage img;
-  if (vkCreateImage(dev.device, &imgCI, nullptr, &img) != VK_SUCCESS)
-  {
-    log->note("Image create failed\n");
-    return -1;
-  }
-
-  // Allocate device-local memory for the image.
-  VkMemoryRequirements imReq;
-  vkGetImageMemoryRequirements(dev.device, img, &imReq);
-  VkPhysicalDeviceMemoryProperties memProps;
-  vkGetPhysicalDeviceMemoryProperties(dev.physicalDevice, &memProps);
-  uint32_t typeIdx = UINT32_MAX;
-  for (uint32_t i = 0; i < memProps.memoryTypeCount; i++)
-    if ((imReq.memoryTypeBits & (1u << i)) &&
-        (memProps.memoryTypes[i].propertyFlags & VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT))
-    { typeIdx = i; break; }
-
-  VkMemoryAllocateInfo aI = {};
-  aI.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-  aI.allocationSize = imReq.size;
-  aI.memoryTypeIndex = typeIdx;
-  VkDeviceMemory imgMem;
-  vkAllocateMemory(dev.device, &aI, nullptr, &imgMem);
-  vkBindImageMemory(dev.device, img, imgMem, 0);
-
-  // Transition UNDEFINED -> SHADER_READ_ONLY_OPTIMAL.  We don't need to
-  // upload any data; the image contents being unspecified is fine for a
-  // bandwidth measurement (the cache lines still get fetched).
-  VkCommandBufferAllocateInfo cbAI = {};
-  cbAI.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-  cbAI.commandPool = dev.commandPool;
-  cbAI.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-  cbAI.commandBufferCount = 1;
-  VkCommandBuffer transCmd;
-  vkAllocateCommandBuffers(dev.device, &cbAI, &transCmd);
-  VkCommandBufferBeginInfo cbBI = {};
-  cbBI.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-  cbBI.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-  vkBeginCommandBuffer(transCmd, &cbBI);
-  VkImageMemoryBarrier b = {};
-  b.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-  b.oldLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-  b.newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-  b.image = img;
-  b.subresourceRange = {VK_IMAGE_ASPECT_COLOR_BIT, 0, 1, 0, 1};
-  b.srcAccessMask = 0;
-  b.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-  vkCmdPipelineBarrier(transCmd, VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
-                       VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, 0,
-                       0, nullptr, 0, nullptr, 1, &b);
-  vkEndCommandBuffer(transCmd);
-  dev.submitAndWait(transCmd);
-  vkFreeCommandBuffers(dev.device, dev.commandPool, 1, &transCmd);
-
-  VkImageViewCreateInfo ivCI = {};
-  ivCI.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-  ivCI.image = img; ivCI.viewType = VK_IMAGE_VIEW_TYPE_2D;
-  ivCI.format = VK_FORMAT_R32G32B32A32_SFLOAT;
-  ivCI.subresourceRange = {VK_IMAGE_ASPECT_COLOR_BIT, 0, 1, 0, 1};
-  VkImageView imgView;
-  vkCreateImageView(dev.device, &ivCI, nullptr, &imgView);
-
-  VkSamplerCreateInfo smCI = {};
-  smCI.sType = VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO;
-  smCI.magFilter = VK_FILTER_NEAREST;
-  smCI.minFilter = VK_FILTER_NEAREST;
-  smCI.mipmapMode = VK_SAMPLER_MIPMAP_MODE_NEAREST;
-  smCI.addressModeU = smCI.addressModeV = smCI.addressModeW =
-      VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE;
-  smCI.unnormalizedCoordinates = VK_FALSE;
-  VkSampler sampler;
-  vkCreateSampler(dev.device, &smCI, nullptr, &sampler);
-
-  VkBuffer outBuf; VkDeviceMemory outMem;
-  if (!dev.createBuffer(outBytes, VK_BUFFER_USAGE_STORAGE_BUFFER_BIT,
-                        VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, outBuf, outMem))
-  {
-    log->note("Output buffer alloc failed\n");
-    vkDestroySampler(dev.device, sampler, nullptr);
-    vkDestroyImageView(dev.device, imgView, nullptr);
-    vkDestroyImage(dev.device, img, nullptr);
-    vkFreeMemory(dev.device, imgMem, nullptr);
-    return -1;
-  }
-
-  VkDescriptorSetLayoutBinding bs[2] = {};
-  bs[0].binding = 0; bs[0].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-  bs[0].descriptorCount = 1; bs[0].stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
-  bs[1].binding = 1; bs[1].descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-  bs[1].descriptorCount = 1; bs[1].stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
-  VkDescriptorSetLayoutCreateInfo dslCI = {};
-  dslCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-  dslCI.bindingCount = 2; dslCI.pBindings = bs;
-  VkDescriptorSetLayout dsLayout;
-  vkCreateDescriptorSetLayout(dev.device, &dslCI, nullptr, &dsLayout);
-
-  VkPipelineLayoutCreateInfo plCI = {};
-  plCI.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-  plCI.setLayoutCount = 1; plCI.pSetLayouts = &dsLayout;
-  VkPipelineLayout pipeLayout;
-  vkCreatePipelineLayout(dev.device, &plCI, nullptr, &pipeLayout);
-
-  VkDescriptorPoolSize ps[2] = {
-    {VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, 1},
-    {VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1},
-  };
-  VkDescriptorPoolCreateInfo dpCI = {};
-  dpCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-  dpCI.maxSets = 1; dpCI.poolSizeCount = 2; dpCI.pPoolSizes = ps;
-  VkDescriptorPool descPool;
-  vkCreateDescriptorPool(dev.device, &dpCI, nullptr, &descPool);
-
-  VkDescriptorSetAllocateInfo dsAI = {};
-  dsAI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-  dsAI.descriptorPool = descPool; dsAI.descriptorSetCount = 1; dsAI.pSetLayouts = &dsLayout;
-  VkDescriptorSet descSet;
-  vkAllocateDescriptorSets(dev.device, &dsAI, &descSet);
-
-  VkDescriptorImageInfo ii = {};
-  ii.imageView = imgView; ii.sampler = sampler;
-  ii.imageLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-  VkDescriptorBufferInfo bi = {outBuf, 0, outBytes};
-  VkWriteDescriptorSet ws[2] = {};
-  ws[0].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-  ws[0].dstSet = descSet; ws[0].dstBinding = 0; ws[0].descriptorCount = 1;
-  ws[0].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-  ws[0].pImageInfo = &ii;
-  ws[1].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-  ws[1].dstSet = descSet; ws[1].dstBinding = 1; ws[1].descriptorCount = 1;
-  ws[1].descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-  ws[1].pBufferInfo = &bi;
-  vkUpdateDescriptorSets(dev.device, 2, ws, 0, nullptr);
-
-  VkPipeline pipe;
-  bool ok = dev.createComputePipeline(vk_shaders::image_bandwidth_v1,
-      vk_shaders::image_bandwidth_v1_size, dsLayout, pipeLayout, pipe);
-  if (!ok)
-  {
-    test.skip("float4", ResultStatus::Error, "Pipeline creation failed");
-  }
-  else
-  {
-    float us = runKernel(dev, pipe, pipeLayout, descSet, numGroups,
-                         cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-    if (us <= 0.0f)
-    {
-      test.skip("float4", ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-    }
-    else
-    {
-      uint64_t bytes = (uint64_t)IMAGE_FETCH_PER_WI * 4 * sizeof(float) * globalWIs;
-      float gbps = (float)bytes / us / 1e3f;
-      test.emit("float4", gbps);
-    }
-    vkDestroyPipeline(dev.device, pipe, nullptr);
-  }
-
-  vkDestroyDescriptorPool(dev.device, descPool, nullptr);
-  vkDestroyPipelineLayout(dev.device, pipeLayout, nullptr);
-  vkDestroyDescriptorSetLayout(dev.device, dsLayout, nullptr);
-  vkDestroyBuffer(dev.device, outBuf, nullptr);
-  vkFreeMemory(dev.device, outMem, nullptr);
-  vkDestroySampler(dev.device, sampler, nullptr);
-  vkDestroyImageView(dev.device, imgView, nullptr);
-  vkDestroyImage(dev.device, img, nullptr);
-  vkFreeMemory(dev.device, imgMem, nullptr);
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// Atomic throughput (Vulkan -- global + local atomics)
-// ---------------------------------------------------------------------------
-
-int vkPeak::runAtomicThroughput(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  logger::TestSpec testSpec;
-  testSpec.tag = "atomic_throughput";
-  testSpec.display = "Atomic throughput";
-  testSpec.unit = "gops";
-  auto test = currentDeviceScope->beginTest(testSpec);
-
-  const uint32_t wgSize = 256;
-  uint64_t globalWIs = targetVulkanGlobalThreads(dev.info);
-  uint32_t numGroups = (uint32_t)(globalWIs / wgSize);
-
-  // Helper: allocate a single-storage-buffer descriptor + dispatch + time.
-  auto runOne = [&](const char *label, const uint32_t *spv, size_t spvSize,
-                    uint64_t bufBytes) -> float
-  {
-    VkBuffer buf; VkDeviceMemory mem;
-    if (!dev.createBuffer(bufBytes,
-                          VK_BUFFER_USAGE_STORAGE_BUFFER_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-                          VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, buf, mem))
-      return -1.0f;
-    if (!dev.zeroBuffer(buf))
-    {
-      vkDestroyBuffer(dev.device, buf, nullptr);
-      vkFreeMemory(dev.device, mem, nullptr);
-      return -1.0f;
-    }
-    VkDescriptorSetLayoutBinding b = {};
-    b.binding = 0; b.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    b.descriptorCount = 1; b.stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
-    VkDescriptorSetLayoutCreateInfo dslCI = {};
-    dslCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-    dslCI.bindingCount = 1; dslCI.pBindings = &b;
-    VkDescriptorSetLayout dsl; vkCreateDescriptorSetLayout(dev.device, &dslCI, nullptr, &dsl);
-    VkPipelineLayoutCreateInfo plCI = {};
-    plCI.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-    plCI.setLayoutCount = 1; plCI.pSetLayouts = &dsl;
-    VkPipelineLayout pl; vkCreatePipelineLayout(dev.device, &plCI, nullptr, &pl);
-    VkDescriptorPoolSize ps = {VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1};
-    VkDescriptorPoolCreateInfo dpCI = {};
-    dpCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-    dpCI.maxSets = 1; dpCI.poolSizeCount = 1; dpCI.pPoolSizes = &ps;
-    VkDescriptorPool dp; vkCreateDescriptorPool(dev.device, &dpCI, nullptr, &dp);
-    VkDescriptorSetAllocateInfo dsAI = {};
-    dsAI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-    dsAI.descriptorPool = dp; dsAI.descriptorSetCount = 1; dsAI.pSetLayouts = &dsl;
-    VkDescriptorSet ds; vkAllocateDescriptorSets(dev.device, &dsAI, &ds);
-    VkDescriptorBufferInfo bi = {buf, 0, bufBytes};
-    VkWriteDescriptorSet w = {};
-    w.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-    w.dstSet = ds; w.dstBinding = 0; w.descriptorCount = 1;
-    w.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER; w.pBufferInfo = &bi;
-    vkUpdateDescriptorSets(dev.device, 1, &w, 0, nullptr);
-
-    VkPipeline pipe;
-    if (!dev.createComputePipeline(spv, spvSize, dsl, pl, pipe))
-    {
-      vkDestroyDescriptorPool(dev.device, dp, nullptr);
-      vkDestroyPipelineLayout(dev.device, pl, nullptr);
-      vkDestroyDescriptorSetLayout(dev.device, dsl, nullptr);
-      vkDestroyBuffer(dev.device, buf, nullptr);
-      vkFreeMemory(dev.device, mem, nullptr);
-      return -1.0f;
-    }
-    float us = runKernel(dev, pipe, pl, ds, numGroups,
-                         cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-    vkDestroyPipeline(dev.device, pipe, nullptr);
-    vkDestroyDescriptorPool(dev.device, dp, nullptr);
-    vkDestroyPipelineLayout(dev.device, pl, nullptr);
-    vkDestroyDescriptorSetLayout(dev.device, dsl, nullptr);
-    vkDestroyBuffer(dev.device, buf, nullptr);
-    vkFreeMemory(dev.device, mem, nullptr);
-    (void)label;
-    return us;
-  };
-
-  // Global atomics: one int counter per WI -> 128 MB output.
-  float us_g = runOne("int_global", vk_shaders::atomic_throughput_global,
-      vk_shaders::atomic_throughput_global_size, globalWIs * sizeof(int32_t));
-  if (us_g > 0)
-  {
-    float gops = ((float)globalWIs * (float)ATOMIC_REPS) / us_g / 1e3f;
-    test.emit("int_global", gops);
-  }
-  else
-  {
-    test.skip("int_global", ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-  }
-
-#ifdef VK_HAS_ATOMIC_THROUGHPUT_GLOBAL_UINT64
-  if (dev.info.atomicInt64Supported)
-  {
-    float us = runOne("ulong_global", vk_shaders::atomic_throughput_global_uint64,
-        vk_shaders::atomic_throughput_global_uint64_size, globalWIs * sizeof(uint64_t));
-    if (us > 0)
-    {
-      float gops = ((float)globalWIs * (float)ATOMIC_REPS) / us / 1e3f;
-      test.emit("ulong_global", gops);
-    }
-    else
-    {
-      test.skip("ulong_global", ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-    }
-  }
-#endif
-
-  // Local atomics: one int counter per workgroup.
-  float us_l = runOne("int_local", vk_shaders::atomic_throughput_local,
-      vk_shaders::atomic_throughput_local_size, (uint64_t)numGroups * sizeof(int32_t));
-  if (us_l > 0)
-  {
-    float gops = ((float)globalWIs * (float)ATOMIC_REPS) / us_l / 1e3f;
-    test.emit("int_local", gops);
-  }
-  else
-  {
-    test.skip("int_local", ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-  }
-
-  return 0;
-}
-
-// ---------------------------------------------------------------------------
-// FP atomic throughput (Vulkan) -- emitted in the FpCompute phase, mirroring
-// Metal's runAtomicThroughputFp.  Currently a single global_float row, gated
-// on VK_EXT_shader_atomic_float + shaderBufferFloat32AtomicAdd.
-// ---------------------------------------------------------------------------
-
-#ifdef VK_HAS_ATOMIC_THROUGHPUT_GLOBAL_FLOAT
-int vkPeak::runAtomicThroughputFp(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  logger::TestSpec testSpec;
-  testSpec.tag = "atomic_throughput";
-  testSpec.display = "Atomic throughput";
-  testSpec.unit = "gflops";
-  auto test = currentDeviceScope->beginTest(testSpec);
-
-  if (!dev.info.atomicFloat32Supported)
-  {
-    test.skip("float_global", ResultStatus::Unsupported,
-              "VK_EXT_shader_atomic_float not supported");
-    return 0;
-  }
-
-  const uint32_t wgSize = 256;
-  uint64_t globalWIs = targetVulkanGlobalThreads(dev.info);
-  uint32_t numGroups = (uint32_t)(globalWIs / wgSize);
-  uint64_t bufBytes = globalWIs * sizeof(float);
-
-  VkBuffer buf; VkDeviceMemory mem;
-  if (!dev.createBuffer(bufBytes,
-                        VK_BUFFER_USAGE_STORAGE_BUFFER_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-                        VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, buf, mem))
-  {
-    log->note("Failed to allocate buffer\n");
-    return -1;
-  }
-  if (!dev.zeroBuffer(buf))
-  {
-    log->note("Failed to zero buffer\n");
-    vkDestroyBuffer(dev.device, buf, nullptr);
-    vkFreeMemory(dev.device, mem, nullptr);
-    return -1;
-  }
-  VkDescriptorSetLayoutBinding b = {};
-  b.binding = 0; b.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-  b.descriptorCount = 1; b.stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
-  VkDescriptorSetLayoutCreateInfo dslCI = {};
-  dslCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-  dslCI.bindingCount = 1; dslCI.pBindings = &b;
-  VkDescriptorSetLayout dsl; vkCreateDescriptorSetLayout(dev.device, &dslCI, nullptr, &dsl);
-  VkPipelineLayoutCreateInfo plCI = {};
-  plCI.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-  plCI.setLayoutCount = 1; plCI.pSetLayouts = &dsl;
-  VkPipelineLayout pl; vkCreatePipelineLayout(dev.device, &plCI, nullptr, &pl);
-  VkDescriptorPoolSize ps = {VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1};
-  VkDescriptorPoolCreateInfo dpCI = {};
-  dpCI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-  dpCI.maxSets = 1; dpCI.poolSizeCount = 1; dpCI.pPoolSizes = &ps;
-  VkDescriptorPool dp; vkCreateDescriptorPool(dev.device, &dpCI, nullptr, &dp);
-  VkDescriptorSetAllocateInfo dsAI = {};
-  dsAI.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-  dsAI.descriptorPool = dp; dsAI.descriptorSetCount = 1; dsAI.pSetLayouts = &dsl;
-  VkDescriptorSet ds; vkAllocateDescriptorSets(dev.device, &dsAI, &ds);
-  VkDescriptorBufferInfo bi = {buf, 0, bufBytes};
-  VkWriteDescriptorSet w = {};
-  w.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-  w.dstSet = ds; w.dstBinding = 0; w.descriptorCount = 1;
-  w.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER; w.pBufferInfo = &bi;
-  vkUpdateDescriptorSets(dev.device, 1, &w, 0, nullptr);
-
-  VkPipeline pipe;
-  if (!dev.createComputePipeline(vk_shaders::atomic_throughput_global_float,
-                                  vk_shaders::atomic_throughput_global_float_size,
-                                  dsl, pl, pipe))
-  {
-    test.skip("float_global", ResultStatus::Error, "Pipeline creation failed");
-  }
-  else
-  {
-    float us = runKernel(dev, pipe, pl, ds, numGroups,
-                         cfg.targetTimeUs, forceIters ? specifiedIters : 0);
-    if (us <= 0.0f)
-    {
-      test.skip("float_global", ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-    }
-    else
-    {
-      float gflops = ((float)globalWIs * (float)ATOMIC_REPS) / us / 1e3f;
-      test.emit("float_global", gflops);
-    }
-    vkDestroyPipeline(dev.device, pipe, nullptr);
-  }
-
-  vkDestroyDescriptorPool(dev.device, dp, nullptr);
-  vkDestroyPipelineLayout(dev.device, pl, nullptr);
-  vkDestroyDescriptorSetLayout(dev.device, dsl, nullptr);
-  vkDestroyBuffer(dev.device, buf, nullptr);
-  vkFreeMemory(dev.device, mem, nullptr);
-  return 0;
-}
-#endif
-
-// ---------------------------------------------------------------------------
-// Kernel launch latency (Vulkan) -- two metrics:
-//
-//   dispatch  : one-way host-submit -> GPU-kernel-start.  Requires
-//               VK_EXT_calibrated_timestamps to map host time into the GPU
-//               clock domain.  Skipped (with note) when the extension is
-//               unavailable.
-//   roundtrip : std::chrono around vkQueueSubmit + vkQueueWaitIdle.  Always
-//               reported, directly comparable to the same metric on
-//               OpenCL/CUDA/Metal.
-// ---------------------------------------------------------------------------
-
-int vkPeak::runKernelLatency(VulkanDevice &dev, benchmark_config_t &cfg)
-{
-  unsigned int iters = forceIters ? specifiedIters
-                                  : (cfg.kernelLatencyIters ? cfg.kernelLatencyIters : 1000);
-
-  logger::TestSpec testSpec;
-  testSpec.tag = "kernel_launch_latency";
-  testSpec.display = "Kernel launch latency";
-  testSpec.unit = "us";
-  auto test = currentDeviceScope->beginTest(testSpec);
-
-  // Pipeline layout with no descriptor sets and no push constants.
-  VkPipelineLayoutCreateInfo plCI = {};
-  plCI.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-  VkPipelineLayout pipeLayout = VK_NULL_HANDLE;
-  if (vkCreatePipelineLayout(dev.device, &plCI, nullptr, &pipeLayout) != VK_SUCCESS)
-  {
-    log->note("pipeline layout creation failed\n");
-    return -1;
-  }
-
-  // Shader module from the embedded no-op SPIR-V.
-  VkShaderModuleCreateInfo smCI = {};
-  smCI.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-  smCI.codeSize = vk_shaders::kernel_latency_size;
-  smCI.pCode    = vk_shaders::kernel_latency;
-  VkShaderModule shaderModule = VK_NULL_HANDLE;
-  if (vkCreateShaderModule(dev.device, &smCI, nullptr, &shaderModule) != VK_SUCCESS)
-  {
-    vkDestroyPipelineLayout(dev.device, pipeLayout, nullptr);
-    log->note("shader module creation failed\n");
-    return -1;
-  }
-
-  VkComputePipelineCreateInfo pCI = {};
-  pCI.sType = VK_STRUCTURE_TYPE_COMPUTE_PIPELINE_CREATE_INFO;
-  pCI.layout = pipeLayout;
-  pCI.stage.sType  = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-  pCI.stage.stage  = VK_SHADER_STAGE_COMPUTE_BIT;
-  pCI.stage.module = shaderModule;
-  pCI.stage.pName  = "main";
-  VkPipeline pipeline = VK_NULL_HANDLE;
-  if (vkCreateComputePipelines(dev.device, VK_NULL_HANDLE, 1, &pCI, nullptr, &pipeline) != VK_SUCCESS)
-  {
-    vkDestroyShaderModule(dev.device, shaderModule, nullptr);
-    vkDestroyPipelineLayout(dev.device, pipeLayout, nullptr);
-    log->note("compute pipeline creation failed\n");
-    return -1;
-  }
-
-  // GPU timestamp query pool used by the dispatch-latency path: a single
-  // TOP_OF_PIPE timestamp tells us when the GPU started executing.
-  VkPhysicalDeviceProperties props;
-  vkGetPhysicalDeviceProperties(dev.physicalDevice, &props);
-  float timestampPeriodNs = props.limits.timestampPeriod;
-  bool haveTimestamps = (timestampPeriodNs > 0.0f) && dev.info.calibratedTimestampsSupported;
-
-  // Resolve VK_EXT_calibrated_timestamps function pointers.  These let us
-  // capture a (host_t, gpu_t) pair "at the same instant" so we can convert
-  // GPU timestamps into the host clock domain and measure exact one-way
-  // dispatch latency, the same way OpenCL profiling reports QUEUED -> START.
-  PFN_vkGetCalibratedTimestampsEXT pfnGetCalTs = nullptr;
-  VkTimeDomainEXT hostDomain = VK_TIME_DOMAIN_DEVICE_EXT; // sentinel: unset
-  if (haveTimestamps)
-  {
-    pfnGetCalTs = (PFN_vkGetCalibratedTimestampsEXT)
-        vkGetDeviceProcAddr(dev.device, "vkGetCalibratedTimestampsEXT");
-    auto pfnGetDomains = (PFN_vkGetPhysicalDeviceCalibrateableTimeDomainsEXT)
-        vkGetInstanceProcAddr(instance, "vkGetPhysicalDeviceCalibrateableTimeDomainsEXT");
-    if (pfnGetCalTs && pfnGetDomains)
-    {
-      uint32_t n = 0;
-      pfnGetDomains(dev.physicalDevice, &n, nullptr);
-      std::vector<VkTimeDomainEXT> domains(n);
-      pfnGetDomains(dev.physicalDevice, &n, domains.data());
-      // Prefer a host monotonic clock so we can read it cheaply with
-      // std::chrono::steady_clock (CLOCK_MONOTONIC on POSIX, QPC on Windows).
-      VkTimeDomainEXT preferred[] = {
-#if defined(_WIN32)
-        VK_TIME_DOMAIN_QUERY_PERFORMANCE_COUNTER_EXT,
-#else
-        VK_TIME_DOMAIN_CLOCK_MONOTONIC_EXT,
-        VK_TIME_DOMAIN_CLOCK_MONOTONIC_RAW_EXT,
-#endif
-      };
-      for (auto p : preferred)
-      {
-        for (auto d : domains) if (d == p) { hostDomain = p; break; }
-        if (hostDomain != VK_TIME_DOMAIN_DEVICE_EXT) break;
-      }
-    }
-    if (hostDomain == VK_TIME_DOMAIN_DEVICE_EXT)
-      haveTimestamps = false; // can't pin host<->device clocks; skip dispatch
-  }
-
-  VkQueryPool queryPool = VK_NULL_HANDLE;
-  if (haveTimestamps)
-  {
-    VkQueryPoolCreateInfo qpCI = {};
-    qpCI.sType = VK_STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO;
-    qpCI.queryType = VK_QUERY_TYPE_TIMESTAMP;
-    qpCI.queryCount = 1;
-    if (vkCreateQueryPool(dev.device, &qpCI, nullptr, &queryPool) != VK_SUCCESS)
-      haveTimestamps = false;
-  }
-
-  // Pre-record a 1x1x1 dispatch.  Reusing one command buffer across iters
-  // keeps record overhead out of the timing loop.
-  VkCommandBufferAllocateInfo allocInfo = {};
-  allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-  allocInfo.commandPool = dev.commandPool;
-  allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-  allocInfo.commandBufferCount = 1;
-  VkCommandBuffer cmdBuf = VK_NULL_HANDLE;
-  vkAllocateCommandBuffers(dev.device, &allocInfo, &cmdBuf);
-
-  VkCommandBufferBeginInfo beginInfo = {};
-  beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-  beginInfo.flags = VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT;
-  vkBeginCommandBuffer(cmdBuf, &beginInfo);
-  if (haveTimestamps)
-  {
-    vkCmdResetQueryPool(cmdBuf, queryPool, 0, 1);
-    vkCmdWriteTimestamp(cmdBuf, VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT, queryPool, 0);
-  }
-  vkCmdBindPipeline(cmdBuf, VK_PIPELINE_BIND_POINT_COMPUTE, pipeline);
-  vkCmdDispatch(cmdBuf, 1, 1, 1);
-  vkEndCommandBuffer(cmdBuf);
-
-  // Get a calibrated (host, gpu) pair.  Stored in driver-domain units:
-  //   ns since CLOCK_MONOTONIC epoch (POSIX) or QPC ticks (Windows).
-  // We re-read the matching host timestamp in the SAME domain inside the
-  // hot loop, so the units always match across the subtraction.
-  auto getCalibrated = [&](uint64_t &hostOut, uint64_t &gpuOut) -> bool {
-    VkCalibratedTimestampInfoEXT infos[2] = {};
-    infos[0].sType = VK_STRUCTURE_TYPE_CALIBRATED_TIMESTAMP_INFO_EXT;
-    infos[0].timeDomain = hostDomain;
-    infos[1].sType = VK_STRUCTURE_TYPE_CALIBRATED_TIMESTAMP_INFO_EXT;
-    infos[1].timeDomain = VK_TIME_DOMAIN_DEVICE_EXT;
-    uint64_t ts[2] = {0, 0};
-    uint64_t maxDev = 0;
-    if (pfnGetCalTs(dev.device, 2, infos, ts, &maxDev) != VK_SUCCESS)
-      return false;
-    hostOut = ts[0];
-    gpuOut  = ts[1];
-    return true;
-  };
-
-  // Read the host clock in the same domain as `hostDomain`.  POSIX:
-  // CLOCK_MONOTONIC nanoseconds; Windows: QPC ticks.  Matches what the
-  // driver returns for VK_TIME_DOMAIN_CLOCK_MONOTONIC{,_RAW}_EXT and
-  // VK_TIME_DOMAIN_QUERY_PERFORMANCE_COUNTER_EXT respectively, so the
-  // raw uint64 can be subtracted directly from `hostCalib`.
-  auto hostNowInDriverDomain = [&]() -> uint64_t {
-#if defined(_WIN32)
-    LARGE_INTEGER c;
-    QueryPerformanceCounter(&c);
-    return (uint64_t)c.QuadPart;
-#else
-    struct timespec ts;
-    clock_gettime((hostDomain == VK_TIME_DOMAIN_CLOCK_MONOTONIC_RAW_EXT)
-                    ? CLOCK_MONOTONIC_RAW : CLOCK_MONOTONIC, &ts);
-    return (uint64_t)ts.tv_sec * 1000000000ULL + (uint64_t)ts.tv_nsec;
-#endif
-  };
-
-  // Conversion factor from host driver-domain "ticks" to nanoseconds.
-  // POSIX: ticks ARE nanoseconds, factor = 1.  Windows: ticks are QPC,
-  // factor = 1e9 / QPC_frequency.
-  double hostTickNs = 1.0;
-#if defined(_WIN32)
-  {
-    LARGE_INTEGER freq;
-    QueryPerformanceFrequency(&freq);
-    hostTickNs = 1e9 / (double)freq.QuadPart;
-  }
-#endif
-
-  // Calibrate ONCE before the loop.  vkGetCalibratedTimestampsEXT itself
-  // costs a few microseconds; doing it per-iter pollutes the dispatch
-  // measurement by exactly that cost.  Clock drift over our short
-  // measurement window is negligible (<<1us across thousands of iters
-  // on every modern OS clock).
-  uint64_t hostCalib = 0, gpuCalib = 0;
-  if (haveTimestamps && !getCalibrated(hostCalib, gpuCalib))
-    haveTimestamps = false;
-
-  bool submitFailed = false;
-
-  // Warmup
-  for (unsigned int w = 0; w < warmupCount; w++)
-  {
-    if (dev.submitAndWait(cmdBuf) != VK_SUCCESS) { submitFailed = true; break; }
-  }
-
-  double totalDispatchUs  = 0;
-  double totalRoundtripUs = 0;
-  unsigned int dispatchSamples = 0;
-  if (!submitFailed)
-  {
-    for (unsigned int i = 0; i < iters; i++)
-    {
-      uint64_t hostSubmit = haveTimestamps ? hostNowInDriverDomain() : 0;
-      auto chronoStart = std::chrono::high_resolution_clock::now();
-      VkResult sr = dev.submitAndWait(cmdBuf);
-      auto chronoEnd = std::chrono::high_resolution_clock::now();
-      if (sr != VK_SUCCESS) { submitFailed = true; break; }
-      totalRoundtripUs += (double)std::chrono::duration_cast<std::chrono::nanoseconds>(chronoEnd - chronoStart).count() / 1000.0;
-
-      if (haveTimestamps)
-      {
-        uint64_t gpuStart = 0;
-        if (vkGetQueryPoolResults(dev.device, queryPool, 0, 1,
-                                  sizeof(gpuStart), &gpuStart, sizeof(uint64_t),
-                                  VK_QUERY_RESULT_64_BIT | VK_QUERY_RESULT_WAIT_BIT) == VK_SUCCESS)
-        {
-          // Convert everything to nanoseconds, then to microseconds.
-          double hostSubmitNs = (double)hostSubmit * hostTickNs;
-          double hostCalibNs  = (double)hostCalib  * hostTickNs;
-          double gpuDeltaNs   = (double)((int64_t)gpuStart - (int64_t)gpuCalib) * (double)timestampPeriodNs;
-          double hostStartNs  = hostCalibNs + gpuDeltaNs;
-          double dispatchUs   = (hostStartNs - hostSubmitNs) / 1000.0;
-          if (dispatchUs > 0)
-          {
-            totalDispatchUs += dispatchUs;
-            dispatchSamples++;
-          }
-        }
-      }
-    }
-  }
-
-  if (submitFailed)
-  {
-    test.skip("dispatch", ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-  }
-  else if (dispatchSamples > 0)
-  {
-    float dispatchUs = (float)(totalDispatchUs / dispatchSamples);
-    test.emit("dispatch", dispatchUs);
-  }
-  else
-  {
-    test.skip("dispatch", ResultStatus::Unsupported,
-               "Needs VK_EXT_calibrated_timestamps");
-  }
-  if (submitFailed)
-  {
-    test.skip("roundtrip", ResultStatus::Error, "vkQueueSubmit/WaitIdle failed");
-  }
-  else
-  {
-    double avgRoundtripUs = totalRoundtripUs / static_cast<double>(iters);
-    test.emit("roundtrip", (float)avgRoundtripUs);
-  }
-
-  vkFreeCommandBuffers(dev.device, dev.commandPool, 1, &cmdBuf);
-  if (queryPool != VK_NULL_HANDLE)
-    vkDestroyQueryPool(dev.device, queryPool, nullptr);
-  vkDestroyPipeline(dev.device, pipeline, nullptr);
-  vkDestroyShaderModule(dev.device, shaderModule, nullptr);
-  vkDestroyPipelineLayout(dev.device, pipeLayout, nullptr);
-
-  return 0;
-}
 
 // Free-function enumeration used by --list-devices (desktop) and the Android
 // JNI surface. Spins up a throwaway VkInstance just long enough to read
